Split: TestPeptideWithSetMods
