--- mzLib/Test/TestPeptideWithSetMods.cs
@@ -1,1697 +1,0 @@
-﻿using Chemistry;
-using NUnit.Framework;
-using Assert = NUnit.Framework.Legacy.ClassicAssert;
-using CollectionAssert = NUnit.Framework.Legacy.CollectionAssert;
-using Proteomics;
-using Proteomics.ProteolyticDigestion;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using MzLibUtil;
-using Omics;
-using Omics.BioPolymer;
-using Omics.Digestion;
-using Omics.Fragmentation;
-using Omics.Modifications;
-using Omics.SequenceConversion;
-using Transcriptomics.Digestion;
-using UsefulProteomicsDatabases;
-using Stopwatch = System.Diagnostics.Stopwatch;
-using Transcriptomics;
-
-namespace Test
-{
-    [TestFixture]
-    [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
-    public static class TestPeptideWithSetMods
-    {
-        private static Stopwatch Stopwatch { get; set; }
-
-        [SetUp]
-        public static void Setuppp()
-        {
-            Stopwatch = new Stopwatch();
-            Stopwatch.Start();
-        }
-
-        [TearDown]
-        public static void TearDown()
-        {
-            Console.WriteLine($"Analysis time: {Stopwatch.Elapsed.Hours}h {Stopwatch.Elapsed.Minutes}m {Stopwatch.Elapsed.Seconds}s");
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests that peptides from different proteases are NOT equal even with identical sequences.
-        /// This is essential for multiprotease parsimony in MetaMorpheus - without this distinction,
-        /// peptides from different enzyme digests would be incorrectly collapsed during protein inference.
-        /// </summary>
-        [Test]
-        public static void TestDifferentProteaseEquals()
-        {
-            Protein myProtein = new Protein("SEQUENCEK", "accession");
-
-            DigestionParams digest1 = new DigestionParams(protease: "trypsin", maxMissedCleavages: 0, initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-            DigestionParams digest2 = new DigestionParams(protease: "Lys-C|P", maxMissedCleavages: 0, initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-
-            PeptideWithSetModifications pep1 = myProtein.Digest(digest1, new List<Modification>(), new List<Modification>()).First();
-            PeptideWithSetModifications pep2 = myProtein.Digest(digest2, new List<Modification>(), new List<Modification>()).First();
-
-            Assert.That(pep1.FullSequence.Equals(pep2.FullSequence));
-            Assert.That(pep1.Parent.Equals(pep2.Parent));
-            Assert.That(!pep1.DigestionParams.DigestionAgent.Equals(pep2.DigestionParams.DigestionAgent));
-            Assert.That(!pep1.Equals(pep2));
-            Assert.That(!pep1.Equals((object)pep2));
-            Assert.That(!pep1.GetHashCode().Equals(pep2.GetHashCode()));
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests type safety between PeptideWithSetModifications and OligoWithSetMods.
-        /// Ensures these different biopolymer types are never incorrectly compared as equal,
-        /// which would corrupt search results in multi-omics analyses.
-        /// </summary>
-        [Test]
-        public static void TestPeptideOligoEquality()
-        {
-            var oligo = new OligoWithSetMods("GUACUG", []);
-            var peptide = new PeptideWithSetModifications("PEPTIDE", []);
-
-            Assert.That(!oligo.Equals(peptide));
-            Assert.That(!peptide.Equals(oligo));
-            Assert.That(!((IBioPolymerWithSetMods)oligo).Equals(peptide));
-            Assert.That(!((IBioPolymerWithSetMods)peptide).Equals(oligo));
-            Assert.That(!((object)oligo).Equals(peptide));
-            Assert.That(!((object)peptide).Equals(oligo));
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests semi-specific digestion doesn't crash with various protein sequences.
-        /// Semi-specific search is essential for identifying endogenous peptides and degradation
-        /// products. This test guards against edge cases that could crash the digestion algorithm.
-        /// </summary>
-        [Test]
-        public static void TestSemiFewCleavages()
-        {
-            Protein protein = new("MQLLRCFSIFSVIASVLAQELTTICEQIPSPTLESTPYSLSTTTILANGKAMQGVFEYYKSVTFVSNCGSHPSTTSKGSPINTQYVF", "P32781");
-            DigestionParams nParams = new("trypsin", 2, 7, 50, searchModeType: CleavageSpecificity.Semi, fragmentationTerminus: FragmentationTerminus.N);
-            DigestionParams cParams = new("trypsin", 2, 7, 50, searchModeType: CleavageSpecificity.Semi, fragmentationTerminus: FragmentationTerminus.C);
-
-            //Unit test is not crashing
-            protein.Digest(nParams, null, null).ToList();
-            protein.Digest(cParams, null, null).ToList();
-
-            Protein protein2 = new Protein("MQFSTVASVAFVALANFVAAESAAAISQITDGQIQATTTATTEATTTAAPSSTVETVSPSSTETISQQTENGAAKAAVGMGAGALAAAAMLL", "P43497");
-            protein2.Digest(nParams, null, null).ToList();
-            protein2.Digest(cParams, null, null).ToList();
-
-            List<TruncationProduct> proteolysisProducts = new List<TruncationProduct>
-            {
-                new TruncationProduct(5, 25, "asdf")
-            };
-
-            //speedy
-            Protein protein3 = new Protein("MQFSTVASVAFVALANFVAAESAAAISQITDGQIQATTTATTEATTTAAPSSTVETVSPSSTETISQQTENGAAKAAVGMGAGALAAAAMLL", "P43497", proteolysisProducts: proteolysisProducts);
-            protein3.Digest(nParams, null, null).ToList();
-            protein3.Digest(cParams, null, null).ToList();
-            cParams = new DigestionParams("trypsin", 0, 7, 9, searchModeType: CleavageSpecificity.Semi, fragmentationTerminus: FragmentationTerminus.C, initiatorMethionineBehavior: InitiatorMethionineBehavior.Cleave);
-            protein3.Digest(cParams, null, null).ToList();
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests that MaxLength constraints are respected in semi/non-specific digestion.
-        /// Without proper length enforcement, search space would explode and include peptides
-        /// that cannot physically be detected, wasting computational resources and increasing FDR.
-        /// </summary>
-        [Test]
-        public static void TestSpeedyNonAndSemiSpecificMaxLength()
-        {
-            Protein Q07065 = new Protein("MPSAKQRGSKGGHGAASPSEKGAHPSGGADDV" +
-                "AKKPPPAPQQPPPPPAPHPQQHPQQHPQNQAHGKGGHRGGGGGGGKSSSSSSASAAAAAA" +
-                "AASSSASCSRRLGRALNFLFYLALVAAAAFSGWCVHHVLEEVQQVRRSHQDFSRQREELGQ" +
-                "GLQGVEQKVQSLQATFGTFESILRSSQHKQDLTEKAVKQGESEVSRISEVLQKLQNEILKDL" +
-                "SDGIHVVKDARERDFTSLENTVEERLTELTKSINDNIAIFTEVQKRSQKEINDMKAKVASLEE" +
-                "SEGNKQDLKALKEAVKEIQTSAKSREWDMEALRSTLQTMESDIYTEVRELVSLKQEQQAFKEA" +
-                "ADTERLALQALTEKLLRSEESVSRLPEEIRRLEEELRQLKSDSHGPKEDGGFRHSEAFEALQQK" +
-                "SQGLDSRLQHVEDGVLSMQVASARQTESLESLLSKSQEHEQRLAALQGRLEGLGSSEADQDGLAST" +
-                "VRSLGETQLVLYGDVEELKRSVGELPSTVESLQKVQEQVHTLLSQDQAQAARLPPQDFLDRLSSLD" +
-                "NLKASVSQVEADLKMLRTAVDSLVAYSVKIETNENNLESAKGLLDDLRNDLDRLFVKVEKIHEKV", "Q07065");
-
-            //Semi
-            DigestionParams semiNParams = new DigestionParams("Asp-N", 3, 7, 50, searchModeType: CleavageSpecificity.Semi, fragmentationTerminus: FragmentationTerminus.N);
-            DigestionParams semiCParams = new DigestionParams("Asp-N", 3, 7, 50, searchModeType: CleavageSpecificity.Semi, fragmentationTerminus: FragmentationTerminus.C);
-            List<PeptideWithSetModifications> nPwsms = Q07065.Digest(semiNParams, null, null).ToList();
-            List<PeptideWithSetModifications> cPwsms = Q07065.Digest(semiCParams, null, null).ToList();
-            Assert.IsFalse(nPwsms.Any(x => x.Length > semiNParams.MaxLength));
-            Assert.IsFalse(cPwsms.Any(x => x.Length > semiCParams.MaxLength));
-            Assert.IsTrue(nPwsms.Any(x => x.Length == semiNParams.MaxPeptideLength));
-            Assert.IsTrue(cPwsms.Any(x => x.Length == semiCParams.MaxPeptideLength));
-
-            //Non
-            DigestionParams nonNParams = new DigestionParams("Asp-N", 20, 7, 50, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.N); //more missed cleavages here so we can test the end
-            DigestionParams nonCParams = new DigestionParams("Asp-N", 3, 7, 50, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.C);
-            nPwsms = Q07065.Digest(nonNParams, null, null).ToList();
-            cPwsms = Q07065.Digest(nonCParams, null, null).ToList();
-            Assert.IsFalse(nPwsms.Any(x => x.Length > nonNParams.MaxLength));
-            Assert.IsFalse(cPwsms.Any(x => x.Length > nonCParams.MaxLength));
-            Assert.IsTrue(nPwsms.Any(x => x.Length == nonNParams.MaxLength));
-            Assert.IsTrue(cPwsms.Any(x => x.Length == nonCParams.MaxLength));
-            Assert.IsTrue(nPwsms.Any(x => x.Length == nonNParams.MinPeptideLength));
-            Assert.IsTrue(cPwsms.Any(x => x.Length == nonCParams.MinPeptideLength));
-        }
-
-        /// <summary>
-        /// CRITICAL: Comprehensive test for non-specific and semi-specific digestion correctness.
-        /// Tests peptide generation, cleavage specificity labeling, duplicate prevention, and
-        /// methionine behavior. Essential for all non-tryptic and semi-tryptic searches.
-        /// </summary>
-        [Test]
-        public static void TestNonAndSemiSpecificDigests()
-        {
-            Protein fiveCleavages = new Protein("MAAKCCKDDKEEKFFKGG", "fiveCleavages"); //protein with 5 K's
-            List<Tuple<string, FragmentationTerminus>> trypticSequencesInducingClevage = new List<Tuple<string, FragmentationTerminus>>
-            {
-                new Tuple<string, FragmentationTerminus>("K",FragmentationTerminus.C ) //cleave at C terminus of K
-            };
-            List<Tuple<string, FragmentationTerminus>> trypticSequencesPreventingClevage = new List<Tuple<string, FragmentationTerminus>>();
-
-            //make two identical proteases, but one is fully specific and one is semi specific
-
-            var motifList = DigestionMotif.ParseDigestionMotifsFromString("K|");
-            Protease trypsinForTestNonAndSemiSpecificDigests = new Protease("trypsinForTestNonAndSemiSpecificDigests", CleavageSpecificity.Full, "asdf", "asdf", motifList);
-            Protease semiTrypsinForTestNonAndSemiSpecificDigests = new Protease("semitrypsinForTestNonAndSemiSpecificDigests", CleavageSpecificity.Semi, "asdf", "asdf", motifList);
-
-            //add these made up proteases to the dictionary
-            ProteaseDictionary.Dictionary.Add(trypsinForTestNonAndSemiSpecificDigests.Name, trypsinForTestNonAndSemiSpecificDigests);
-            ProteaseDictionary.Dictionary.Add(semiTrypsinForTestNonAndSemiSpecificDigests.Name, semiTrypsinForTestNonAndSemiSpecificDigests);
-
-            //Digest with the full
-            DigestionParams fullyDigestParams = new DigestionParams(trypsinForTestNonAndSemiSpecificDigests.Name, 3, 2);
-            var fiveCleavageProductsTrypsin = fiveCleavages.Digest(fullyDigestParams, null, null).ToList();
-            Assert.AreEqual(22, fiveCleavageProductsTrypsin.Count);
-
-            //digests with the semi (variable methionine)
-            DigestionParams semiDigestionParams = new DigestionParams(semiTrypsinForTestNonAndSemiSpecificDigests.Name, 3, 2);
-            var fiveCleavageProductsSemiTrypsin = fiveCleavages.Digest(semiDigestionParams, null, null).ToList();
-
-            //This is a partial list of the full peptides. From this, we can GENERATE every possible semi that we would expect to see
-            List<string> expectedProductsSemiFiveCleavages = new List<string>
-            {
-                "MAAKCCKDDKEEK",
-                "AAKCCKDDKEEK",
-                "CCKDDKEEKFFK",
-                "DDKEEKFFKGG",
-                "EEKFFKGG",
-                "FFKGG",
-                "AAK",
-                "AAKCCK",
-                "AAKCCKDDK",
-            };
-
-            //Check that, when we digested with semi, we made all possible semi sequences, labeled full and semi correctly, and have no duplicates
-            foreach (string s in expectedProductsSemiFiveCleavages) //foreach precursor peptide
-            {
-                for (int i = 0; i < s.Length - semiDigestionParams.MinLength; i++) //cleave it to be semi
-                {
-                    string sToFind = s.Substring(i); //get a peptide from this precursor (fixed C)
-                    var peps = fiveCleavageProductsSemiTrypsin.Where(x => x.BaseSequence.Equals(sToFind)).ToArray(); //find the peptide in the digested list
-                    Assert.IsTrue(peps.Length == 1); //There should be exactly one! More than that means there are duplicates, fewer means we didn't generate it!
-                    var pep = peps[0]; //get that single peptide
-                    //if it's a full sequence (cleaved at both indexes (including termini))
-                    if ((pep.BaseSequence[0] == pep.BaseSequence[1] || pep.BaseSequence[0] == 'M') && (pep.BaseSequence[pep.BaseSequence.Length - 1] == 'K'
-                        || (pep.BaseSequence[pep.BaseSequence.Length - 1] == 'G' && pep.BaseSequence[pep.BaseSequence.Length - 2] == 'G')))
-                    {
-                        Assert.IsTrue(pep.CleavageSpecificityForFdrCategory == CleavageSpecificity.Full);
-                    }
-                    else
-                    {
-                        Assert.IsTrue(pep.CleavageSpecificityForFdrCategory == CleavageSpecificity.Semi);
-                    }
-                    //try to remake the pwsm with unknown specificity... was it assigned the correct specificity?
-                    PeptideWithSetModifications pwsmRemake = new PeptideWithSetModifications(fiveCleavages, semiDigestionParams, pep.OneBasedStartResidueInProtein, pep.OneBasedEndResidueInProtein, CleavageSpecificity.Unknown, "", 3, pep.AllModsOneIsNterminus, 0);
-                    Assert.IsTrue(pwsmRemake.CleavageSpecificityForFdrCategory == pep.CleavageSpecificityForFdrCategory);
-
-                    //Repeat the above going from the other direction (fixed N)
-                    sToFind = s.Substring(0, semiDigestionParams.MinLength + i); //get a peptide from this precursor (fixed N)
-                    peps = fiveCleavageProductsSemiTrypsin.Where(x => x.BaseSequence.Equals(sToFind)).ToArray();//find the peptide in the digested list
-                    Assert.IsTrue(peps.Length == 1);//There should be exactly one! More than that means there are duplicates, fewer means we didn't generate it!
-                    pep = peps[0];//get that single peptide
-                    //if it's a full sequence (cleaved at both indexes (including termini))
-                    if ((pep.BaseSequence[0] == pep.BaseSequence[1] || pep.BaseSequence[0] == 'M') && pep.BaseSequence.Last() == 'K')
-                    {
-                        Assert.IsTrue(pep.CleavageSpecificityForFdrCategory == CleavageSpecificity.Full);
-                    }
-                    else
-                    {
-                        Assert.IsTrue(pep.CleavageSpecificityForFdrCategory == CleavageSpecificity.Semi);
-                    }
-                    //try to remake the pwsm with unknown specificity... was it assigned the correct specificity?
-                    pwsmRemake = new PeptideWithSetModifications(fiveCleavages, semiDigestionParams, pep.OneBasedStartResidueInProtein, pep.OneBasedEndResidueInProtein, CleavageSpecificity.Unknown, "", 3, pep.AllModsOneIsNterminus, 0);
-                    Assert.IsTrue(pwsmRemake.CleavageSpecificityForFdrCategory == pep.CleavageSpecificityForFdrCategory);
-                }
-            }
-            //confirm there were 85 peptides generated by the semi
-            Assert.AreEqual(85, fiveCleavageProductsSemiTrypsin.Count);
-
-            //The rest of the tests are less intense
-            //check semi when methionine is cleaved
-            DigestionParams semiCleaveDigestionParams = new DigestionParams(semiTrypsinForTestNonAndSemiSpecificDigests.Name, 3, 2, initiatorMethionineBehavior: InitiatorMethionineBehavior.Cleave);
-            var fiveCleavageProductsSemiTrypsinCleave = fiveCleavages.Digest(semiCleaveDigestionParams, null, null).ToList();
-            int numVariableWithMet = fiveCleavageProductsSemiTrypsin.Where(x => x.BaseSequence[0] == 'M').Count(); //how many had methionine in the variable digestion?
-            Assert.AreEqual(fiveCleavageProductsSemiTrypsin.Count, fiveCleavageProductsSemiTrypsinCleave.Count + numVariableWithMet); //there should be the same number of sequences as before, minus the amount of methionine peptides
-
-            //check semi when methionine is retained
-            DigestionParams semiRetainDigestionParams = new DigestionParams(semiTrypsinForTestNonAndSemiSpecificDigests.Name, 3, 2, initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-            var fiveCleavageProductsSemiTrypsinRetain = fiveCleavages.Digest(semiRetainDigestionParams, null, null).ToList();
-            int numNotRetained = fiveCleavageProductsSemiTrypsin.Where(x => x.BaseSequence[0] == 'A' && x.BaseSequence[1] == 'A' //how many had Met cleaved in the variable digestion?
-            && (x.BaseSequence[x.BaseSequence.Length - 1] != 'K' && !(x.BaseSequence[x.BaseSequence.Length - 1] == 'G' && x.BaseSequence[x.BaseSequence.Length - 2] == 'G'))).Count();
-            Assert.AreEqual(fiveCleavageProductsSemiTrypsinRetain.Count + numNotRetained, fiveCleavageProductsSemiTrypsin.Count); //there should be the same number of sequences as before, minus the amount of cleaved peptides
-
-            //Check the speedy semi-specific search (the previous ones were the slow classic)
-            //Fixed N
-            DigestionParams modernSemiDigestionParamsN = new DigestionParams(trypsinForTestNonAndSemiSpecificDigests.Name, 3, 2, searchModeType: CleavageSpecificity.Semi, fragmentationTerminus: FragmentationTerminus.N);
-            var fiveCleavageProductsModernSemiTrypsinN = fiveCleavages.Digest(modernSemiDigestionParamsN, null, null).ToList();
-            Assert.AreEqual(7, fiveCleavageProductsModernSemiTrypsinN.Count);
-
-            //Fixed C
-            DigestionParams modernSemiDigestionParamsC = new DigestionParams(trypsinForTestNonAndSemiSpecificDigests.Name, 3, 2, searchModeType: CleavageSpecificity.Semi, fragmentationTerminus: FragmentationTerminus.C);
-            var fiveCleavageProductsModernSemiTrypsinC = fiveCleavages.Digest(modernSemiDigestionParamsC, null, null).ToList();
-            Assert.AreEqual(6, fiveCleavageProductsModernSemiTrypsinC.Count);
-
-            //test the maxPeptideLength for both singleN and SingleC (variable methionine)
-            //Single N max peptide length
-            var modernNonSpecificN = new DigestionParams("singleN", 4, 2, 4, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.N);
-            var fiveCleavageProductsModernNonSpecificN = fiveCleavages.Digest(modernNonSpecificN, null, null).ToList();
-            Assert.AreEqual(17, fiveCleavageProductsModernNonSpecificN.Count);
-            foreach (var pep in fiveCleavageProductsModernNonSpecificN)
-            {
-                Assert.IsTrue(pep.BaseSequence.Length <= 4 && pep.BaseSequence.Length >= 2);
-            }
-
-            //Single C max peptide length
-            var modernNonSpecificC = new DigestionParams("singleC", 4, 2, 4, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.C);
-            var fiveCleavageProductsModernNonSpecificC = fiveCleavages.Digest(modernNonSpecificC, null, null).ToList();
-            Assert.AreEqual(17, fiveCleavageProductsModernNonSpecificC.Count);
-            foreach (var pep in fiveCleavageProductsModernNonSpecificC)
-            {
-                Assert.IsTrue(pep.BaseSequence.Length <= 4 && pep.BaseSequence.Length >= 2);
-            }
-
-            //test speedy nonspecific with variable methionine
-            TestSingleProteases(fiveCleavages, InitiatorMethionineBehavior.Variable, FragmentationTerminus.N, 17);
-            TestSingleProteases(fiveCleavages, InitiatorMethionineBehavior.Variable, FragmentationTerminus.C, 17);
-
-            //test speedy nonspecific with cleaved methionine
-            TestSingleProteases(fiveCleavages, InitiatorMethionineBehavior.Cleave, FragmentationTerminus.N, 16);
-            TestSingleProteases(fiveCleavages, InitiatorMethionineBehavior.Cleave, FragmentationTerminus.C, 16);
-
-            //test speedy nonspecific with retained methionine
-            TestSingleProteases(fiveCleavages, InitiatorMethionineBehavior.Retain, FragmentationTerminus.N, 17);
-            TestSingleProteases(fiveCleavages, InitiatorMethionineBehavior.Retain, FragmentationTerminus.C, 17);
-
-            //test classic nonspecific
-            DigestionParams classicNonspecificDigest = new DigestionParams("non-specific", 50);
-            List<PeptideWithSetModifications> classicNonspecificPeptides = fiveCleavages.Digest(classicNonspecificDigest, null, null).ToList();
-            Assert.IsTrue(classicNonspecificPeptides.Count == 78);
-        }
-
-        private static void TestSingleProteases(Protein protein, InitiatorMethionineBehavior initiatorMethionineBehavior, FragmentationTerminus fragmentationTerminus, int numSequencesExpected)
-        {
-            string protease = FragmentationTerminus.N == fragmentationTerminus ? "singleN" : "singleC";
-            DigestionParams digestionParams = new DigestionParams(protease, 50, 2, searchModeType: CleavageSpecificity.None, initiatorMethionineBehavior: initiatorMethionineBehavior, fragmentationTerminus: fragmentationTerminus);
-            var products = protein.Digest(digestionParams, null, null).ToList();
-            Assert.AreEqual(numSequencesExpected, products.Count);
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests that terminal modifications are correctly handled with single proteases.
-        /// For SingleN/SingleC searches, terminal mods that don't affect fragment ions should be
-        /// excluded to prevent redundant peptide generation. Essential for spectral library searches.
-        /// </summary>
-        [Test]
-        public static void TestSingleProteasesWithTerminalMods()
-        {
-            //we actually don't want C-terminal mods on SingleN or N-terminal mods on SingleC, because they don't influence the fragment ion series and just create a redundant peptide
-            //the modified peptides are found using precursor mass matching after scoring, but that happens downstream in MetaMorpheus
-            ModificationMotif.TryGetMotif("A", out ModificationMotif motif);
-            Modification nTermMod = new Modification(_originalId: "acetylation", _modificationType: "testModType", _target: motif, _chemicalFormula: ChemicalFormula.ParseFormula("C2H2O1"), _locationRestriction: "N-terminal.");
-            Modification cTermMod = new Modification(_originalId: "amide", _modificationType: "testModType", _target: motif, _chemicalFormula: ChemicalFormula.ParseFormula("C2H2O1"), _locationRestriction: "C-terminal.");
-
-            Protein proteinWithMods = new Protein("MAGIAAKLAKDREAAEGLGSHA", "testProtein",
-                oneBasedModifications: new Dictionary<int, List<Modification>>
-                {
-                    { 2, new List<Modification>{nTermMod } },
-                    { 22, new List<Modification>{cTermMod } }
-                });
-
-            DigestionParams singleN = new DigestionParams(protease: "singleN", searchModeType: CleavageSpecificity.SingleN, fragmentationTerminus: FragmentationTerminus.N);
-            DigestionParams singleC = new DigestionParams(protease: "singleC", searchModeType: CleavageSpecificity.SingleC, fragmentationTerminus: FragmentationTerminus.C);
-
-            List<Modification> empty = new List<Modification>();
-            List<Modification> allMods = new List<Modification> { nTermMod, cTermMod };
-            List<PeptideWithSetModifications> nPeps = proteinWithMods.Digest(singleN, empty, empty).ToList();
-            List<PeptideWithSetModifications> cPeps = proteinWithMods.Digest(singleC, empty, empty).ToList();
-            Assert.IsTrue(nPeps.Count == cPeps.Count);
-            Assert.IsTrue(cPeps.Count == 17);
-
-            Protein proteinWithoutMods = new Protein("MAGIAAKLAKDREAAEGLGSHA", "testProtein");
-
-            //Test that variable mods are removed
-            nPeps = proteinWithoutMods.Digest(singleN, empty, allMods).ToList();
-            cPeps = proteinWithoutMods.Digest(singleC, empty, allMods).ToList();
-            Assert.IsTrue(nPeps.Count == cPeps.Count);
-            Assert.IsTrue(cPeps.Count == 17);
-
-            //Test that fixed mods are NOT removed
-            nPeps = proteinWithoutMods.Digest(singleN, allMods, empty).ToList();
-            cPeps = proteinWithoutMods.Digest(singleC, allMods, empty).ToList();
-            Assert.IsTrue(nPeps.Count == cPeps.Count);
-            Assert.IsTrue(nPeps.All(x => x.FullSequence.Contains("testModType:amide on A")));
-            Assert.IsTrue(nPeps.Last().FullSequence.Contains("testModType:amide on A"));
-            Assert.IsTrue(cPeps.Count == 16);
-
-            //Test single proteases with specific protease
-            DigestionParams specificNonN = new DigestionParams(protease: "Asp-N", searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.N);
-            DigestionParams specificNonC = new DigestionParams(protease: "Asp-N", searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.C);
-            List<PeptideWithSetModifications> nSpecificPeps = proteinWithMods.Digest(specificNonN, empty, empty).ToList();
-            List<PeptideWithSetModifications> cSpecificPeps = proteinWithMods.Digest(specificNonC, empty, empty).ToList();
-            Assert.IsTrue(nSpecificPeps.Count == cSpecificPeps.Count);
-            Assert.IsTrue(cSpecificPeps.Count == 17);
-
-            //try again with no missed cleavages
-            specificNonN = new DigestionParams(protease: "Asp-N", maxMissedCleavages: 0, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.N);
-            specificNonC = new DigestionParams(protease: "Asp-N", maxMissedCleavages: 0, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.C);
-            nSpecificPeps = proteinWithMods.Digest(specificNonN, empty, empty).ToList();
-            cSpecificPeps = proteinWithMods.Digest(specificNonC, empty, empty).ToList();
-            Assert.IsTrue(nSpecificPeps.Count == 11);
-            Assert.IsTrue(cSpecificPeps.Count == 11);
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests single proteases (N/C terminus fixed) with specific cleavage enzymes.
-        /// Validates peptide counts and methionine cleavage behavior for non-specific searches
-        /// that still respect enzyme cleavage sites. Essential for hybrid search strategies.
-        /// </summary>
-        [Test]
-        public static void TestSingleProteasesWithSpecificProteases()
-        {
-            Protein tinyProteinWithCleavages = new Protein("ACDREFGHIKLMNPQRST", "tiny");
-            Protein tinyProteinWithoutCleavages = new Protein("ACDEFGHILMNPQST", "tinier");
-            Protein bigProteinWithStretchOfNoCleavages = new Protein("ACDREFGHIKLMNPQRSTGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGGACDREFGHIKLMNPQRST", "big");
-
-            DigestionParams dpN = new DigestionParams("trypsin", 1, 5, 20, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.N);
-            DigestionParams dpC = new DigestionParams("trypsin", 1, 5, 20, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.C);
-            List<Modification> empty = new List<Modification>();
-
-            //SingleN tests
-
-            List<PeptideWithSetModifications> peptides = tinyProteinWithCleavages.Digest(dpN, empty, empty).ToList();
-            Assert.IsTrue(peptides.Count == 14);
-            peptides = tinyProteinWithoutCleavages.Digest(dpN, empty, empty).ToList();
-            Assert.IsTrue(peptides.Count == 11);
-            peptides = bigProteinWithStretchOfNoCleavages.Digest(dpN, empty, empty).ToList();
-            Assert.IsTrue(peptides.Count == 63);
-
-            //SingleC tests
-            peptides = tinyProteinWithCleavages.Digest(dpC, empty, empty).ToList();
-            Assert.IsTrue(peptides.Count == 14);
-            peptides = tinyProteinWithoutCleavages.Digest(dpC, empty, empty).ToList();
-            Assert.IsTrue(peptides.Count == 11);
-            peptides = bigProteinWithStretchOfNoCleavages.Digest(dpC, empty, empty).ToList();
-            Assert.IsTrue(peptides.Count == 63);
-
-            //Methionine cleavage fringe test
-            Protein methionineProtein = new Protein("MDBCEFGDHIKLMNODPQRST", "tiny");
-            dpN = new DigestionParams("Asp-N", 2, 5, 20, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.N, initiatorMethionineBehavior: InitiatorMethionineBehavior.Cleave);
-            dpC = new DigestionParams("Asp-N", 2, 5, 20, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.C, initiatorMethionineBehavior: InitiatorMethionineBehavior.Cleave);
-            peptides = methionineProtein.Digest(dpN, empty, empty).ToList();
-            Assert.IsTrue(peptides.Count == 16);
-            peptides = methionineProtein.Digest(dpC, empty, empty).ToList();
-            Assert.IsTrue(peptides.Count == 16);
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests parsing of complex modification strings with nested brackets.
-        /// Modifications like "Cation:Fe[III]" contain brackets that could break parsing.
-        /// Essential for correctly reading user-specified and database modifications.
-        /// </summary>
-        [Test]
-        public static void TestHardToParseModifiedSequence()
-        {
-            string fullSequence = "PE[Metal:Cation:Fe[III] on X]PTIDE";
-
-            ModificationMotif.TryGetMotif("X", out var motif);
-
-            Modification mod = new Modification(_originalId: "Cation:Fe[III]", _modificationType: "Metal",
-                _monoisotopicMass: 1, _locationRestriction: "Anywhere.", _target: motif);
-
-            Dictionary<string, Modification> mods = new Dictionary<string, Modification> { { "Cation:Fe[III] on X", mod } };
-
-            PeptideWithSetModifications pep = new PeptideWithSetModifications(fullSequence, mods);
-
-            Assert.That(pep.AllModsOneIsNterminus.Count == 1);
-            var annotatedMod = pep.AllModsOneIsNterminus.First();
-            Assert.That(annotatedMod.Key == 3);
-            Assert.That(annotatedMod.Value.IdWithMotif == "Cation:Fe[III] on X");
-            Assert.That(annotatedMod.Value.OriginalId == "Cation:Fe[III]");
-            Assert.That(annotatedMod.Value.ModificationType == "Metal");
-
-            fullSequence = "[Metal:Cation:Fe[III] on X]PE[Metal:Cation:Fe[III] on X]PTIDE[Metal:Cation:Fe[III] on X]";
-            pep = new PeptideWithSetModifications(fullSequence, mods);
-            Assert.That(pep.AllModsOneIsNterminus.Count == 3);
-            Assert.That(pep.AllModsOneIsNterminus.Keys.ToList().SequenceEqual(new int[] { 1, 3, 8 }));
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests correct parsing of C-terminal vs side-chain modifications.
-        /// A modification on the last residue's side chain vs the peptide C-terminus are
-        /// chemically distinct. Incorrect parsing would cause wrong mass calculations.
-        /// </summary>
-        [Test]
-        public static void TestCTermAndLastSideChainModParsing()
-        {
-            string fullSequenceBothMods = "PEPTIDE[Mod:MyMod on E]-[PeptideCTermMod:MyCTermMod on E]";
-            string fullSequenceCTermOnly = "PEPTIDE-[PeptideCTermMod:MyCTermMod on E]";
-            string fullSequenceSideChainOnly = "PEPTIDE[Mod:MyMod on E]";
-
-            ModificationMotif.TryGetMotif("E", out var motif);
-
-            Modification mod = new Modification(_originalId: "MyMod", _modificationType: "Mod",
-                _monoisotopicMass: 1, _locationRestriction: "Anywhere.", _target: motif);
-
-            Modification cTermMod = new Modification(_originalId: "MyCTermMod", _modificationType: "PeptideCTermMod",
-                _monoisotopicMass: 1, _locationRestriction: "Peptide C-terminal.", _target: motif);
-
-            Dictionary<string, Modification> mods = new Dictionary<string, Modification>
-            {
-                { "MyMod on E", mod },
-                { "MyCTermMod on E", cTermMod }
-            };
-
-            PeptideWithSetModifications pepBothMods = new PeptideWithSetModifications(fullSequenceBothMods, mods);
-            PeptideWithSetModifications pepCterm = new PeptideWithSetModifications(fullSequenceCTermOnly, mods);
-            PeptideWithSetModifications pepSideChain = new PeptideWithSetModifications(fullSequenceSideChainOnly, mods);
-
-            Assert.That(pepBothMods.AllModsOneIsNterminus.Count == 2);
-            Assert.That(pepBothMods.AllModsOneIsNterminus.Keys.SequenceEqual(new int[] { 8, 9 }));
-            Assert.That(pepBothMods.AllModsOneIsNterminus[8].IdWithMotif == "MyMod on E");
-            Assert.That(pepBothMods.AllModsOneIsNterminus[9].IdWithMotif == "MyCTermMod on E");
-            Assert.That(pepCterm.AllModsOneIsNterminus.Count == 1);
-            Assert.That(pepCterm.AllModsOneIsNterminus.Keys.SequenceEqual(new int[] { 9 }));
-            Assert.That(pepCterm.AllModsOneIsNterminus[9].IdWithMotif == "MyCTermMod on E");
-            Assert.That(pepSideChain.AllModsOneIsNterminus.Count == 1);
-            Assert.That(pepSideChain.AllModsOneIsNterminus.Keys.SequenceEqual(new int[] { 8 }));
-            Assert.That(pepSideChain.AllModsOneIsNterminus[8].IdWithMotif == "MyMod on E");
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests hash code generation for peptides with and without digestion params.
-        /// Hash codes are essential for Dictionary/HashSet operations. A null return or
-        /// inconsistent hashing would break peptide deduplication and grouping operations.
-        /// </summary>
-        [Test]
-        public static void TestPeptideWithSetMod_GetHashCode()
-        {
-            PeptideWithSetModifications pep1 = new PeptideWithSetModifications("SEQUENCEK", new Dictionary<string, Modification>());
-            int oneHashCode = pep1.GetHashCode();
-
-            //if digestion params is not defined, the peptidewithsetmods should still return a hashcode.
-            Assert.IsNotNull(oneHashCode);
-
-            Protein myProtein = new Protein("SEQUENCEK", "accession");
-
-            DigestionParams digest1 = new DigestionParams(protease: "trypsin", maxMissedCleavages: 0, initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-
-            PeptideWithSetModifications pep2 = myProtein.Digest(digest1, new List<Modification>(), new List<Modification>()).First();
-
-            int twoHashCode = pep2.GetHashCode();
-
-            //if digestion params IS defined, the peptidewithsetmods should  return a hashcode.
-            Assert.IsNotNull(twoHashCode);
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests top-down proteomics digestion with proteolysis products.
-        /// Validates that full-length proteins and their truncation products are correctly
-        /// generated. Essential for intact protein and proteoform identification.
-        /// </summary>
-        [Test]
-        public static void TestTopDownDigestion()
-        {
-            List<TruncationProduct> proteolysisProducts = new List<TruncationProduct>
-            {
-                new TruncationProduct(5, 20, "asdf")
-            };
-            Protein protein = new Protein("MACDEFGHIKLMNOPQRSTVWYMACDEFGHIKLMNOPQRSTVWYMACDEFGHIKLMNOPQRSTVWY", "testProtein", "Mus", proteolysisProducts: proteolysisProducts);
-            DigestionParams topdownParams = new DigestionParams("top-down");
-            List<PeptideWithSetModifications> peptides = protein.Digest(topdownParams, null, null).ToList();
-            Assert.IsTrue(peptides.Count == 3);
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests cleavage specificity assignment for FDR categorization.
-        /// Peptides must be correctly labeled as Full/Semi/None specificity for proper
-        /// FDR estimation. Incorrect labeling leads to inflated or deflated FDR.
-        /// </summary>
-        [Test]
-        public static void TestUpdateCleavageSpecificity()
-        {
-            Protein protein = new Protein("MACDEFGHIKLMNPQRST", "test");
-            DigestionParams dpVariable = new DigestionParams();
-            DigestionParams dpRetain = new DigestionParams(initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-            Dictionary<int, Modification> empty = new Dictionary<int, Modification>();
-
-            //Test with varying Methionine
-            PeptideWithSetModifications fullCleavageVariableMet = new PeptideWithSetModifications(protein, dpVariable, 2, 10, CleavageSpecificity.Unknown, "", 0, empty, 0);
-            Assert.IsTrue(fullCleavageVariableMet.CleavageSpecificityForFdrCategory == CleavageSpecificity.Full);
-            PeptideWithSetModifications fullCleavageRetainMet = new PeptideWithSetModifications(protein, dpRetain, 2, 10, CleavageSpecificity.Unknown, "", 0, empty, 0);
-            Assert.IsTrue(fullCleavageRetainMet.CleavageSpecificityForFdrCategory == CleavageSpecificity.Semi);
-            PeptideWithSetModifications semiCleavageVariableMet = new PeptideWithSetModifications(protein, dpVariable, 2, 9, CleavageSpecificity.Unknown, "", 0, empty, 0);
-            Assert.IsTrue(semiCleavageVariableMet.CleavageSpecificityForFdrCategory == CleavageSpecificity.Semi);
-            PeptideWithSetModifications semiCleavageRetainMet = new PeptideWithSetModifications(protein, dpRetain, 2, 9, CleavageSpecificity.Unknown, "", 0, empty, 0);
-            Assert.IsTrue(semiCleavageRetainMet.CleavageSpecificityForFdrCategory == CleavageSpecificity.None);
-            PeptideWithSetModifications noneCleavageVariableMet = new PeptideWithSetModifications(protein, dpVariable, 3, 9, CleavageSpecificity.Unknown, "", 0, empty, 0);
-            Assert.IsTrue(noneCleavageVariableMet.CleavageSpecificityForFdrCategory == CleavageSpecificity.None);
-            PeptideWithSetModifications noneCleavageRetainMet = new PeptideWithSetModifications(protein, dpRetain, 3, 9, CleavageSpecificity.Unknown, "", 0, empty, 0);
-            Assert.IsTrue(noneCleavageRetainMet.CleavageSpecificityForFdrCategory == CleavageSpecificity.None);
-
-            //Test with proteolytic cleavages
-            protein = new Protein("MACDEFGHIKLMNPQRST", "test", proteolysisProducts: new List<TruncationProduct> { new TruncationProduct(3, 9, "chain") });
-            PeptideWithSetModifications fullProteolytic = new PeptideWithSetModifications(protein, dpVariable, 3, 9, CleavageSpecificity.Unknown, "", 0, empty, 0);
-            Assert.IsTrue(fullProteolytic.CleavageSpecificityForFdrCategory == CleavageSpecificity.Full);
-            fullProteolytic = new PeptideWithSetModifications(protein, dpVariable, 3, 10, CleavageSpecificity.Unknown, "", 0, empty, 0);
-            Assert.IsTrue(fullProteolytic.CleavageSpecificityForFdrCategory == CleavageSpecificity.Full);
-            PeptideWithSetModifications semiProteolytic = new PeptideWithSetModifications(protein, dpVariable, 3, 6, CleavageSpecificity.Unknown, "", 0, empty, 0);
-            Assert.IsTrue(semiProteolytic.CleavageSpecificityForFdrCategory == CleavageSpecificity.Semi);
-            semiProteolytic = new PeptideWithSetModifications(protein, dpVariable, 5, 9, CleavageSpecificity.Unknown, "", 0, empty, 0);
-            Assert.IsTrue(semiProteolytic.CleavageSpecificityForFdrCategory == CleavageSpecificity.Semi);
-        }
-
-        /// <summary>
-        /// EDGE CASE: Tests single proteases on small proteins where peptide count
-        /// should equal protein length minus minimum peptide length plus one.
-        /// Guards against off-by-one errors in peptide enumeration.
-        /// </summary>
-        [Test]
-        public static void TestSingleProteasesTinyProtein()
-        {
-            Protein P56381 = new Protein("MVAYWRQAGLSYIRYSQICAKAVRDALKTEFKANAEKTSGSNVKIVKVKKE", "P56381");
-            DigestionParams singleN = new DigestionParams(protease: "Asp-N", maxMissedCleavages: 3, minPeptideLength: 7, maxPeptideLength: 50, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.N);
-            DigestionParams singleC = new DigestionParams(protease: "Asp-N", maxMissedCleavages: 3, minPeptideLength: 7, maxPeptideLength: 50, searchModeType: CleavageSpecificity.None, fragmentationTerminus: FragmentationTerminus.C);
-            List<PeptideWithSetModifications> nPwsms = P56381.Digest(singleN, null, null).ToList();
-            List<PeptideWithSetModifications> cPwsms = P56381.Digest(singleC, null, null).ToList();
-            Assert.IsTrue(nPwsms.Count == cPwsms.Count);
-            Assert.IsTrue(nPwsms.Count == P56381.Length - singleN.MinLength + 1);
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests splice site intersection logic for variant peptide analysis.
-        /// Correctly identifying which peptides span splice sites is essential for
-        /// alternative splicing detection and isoform-specific peptide annotation.
-        /// </summary>
-        [Test]
-        public static void TestIncludeSpliceSiteRanges()
-        {
-            Protein protein = new Protein("MACDEFGHIKLMNPQRST", "test");
-            PeptideWithSetModifications pepe = new PeptideWithSetModifications(protein, new DigestionParams(), 2, 10, CleavageSpecificity.Unknown, "", 0, new Dictionary<int, Modification>(), 0);
-            SpliceSite ss1Before = new SpliceSite(1, 1, "");
-            SpliceSite ss2BeginningBefore = new SpliceSite(1, 2, "");
-            SpliceSite ss3 = new SpliceSite(2, 3, "");
-            SpliceSite ss4 = new SpliceSite(3, 4, "");
-            SpliceSite ss5 = new SpliceSite(9, 10, "");
-            SpliceSite ss6EndAfter = new SpliceSite(10, 11, "");
-            SpliceSite ss7After = new SpliceSite(11, 12, "");
-            Assert.IsFalse(pepe.IncludesSpliceSite(ss1Before));
-            Assert.IsFalse(pepe.IncludesSpliceSite(ss2BeginningBefore));
-            Assert.IsTrue(pepe.IncludesSpliceSite(ss3));
-            Assert.IsTrue(pepe.IncludesSpliceSite(ss4));
-            Assert.IsTrue(pepe.IncludesSpliceSite(ss5));
-            Assert.IsFalse(pepe.IncludesSpliceSite(ss6EndAfter));
-            Assert.IsFalse(pepe.IncludesSpliceSite(ss7After));
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests sequence variation intersection and identification logic.
-        /// Validates both 'intersects' (peptide overlaps variant) and 'identifies'
-        /// (peptide uniquely confirms variant) for SAV, MNV, indels, and truncations.
-        /// Essential for variant peptide identification in proteogenomics.
-        /// </summary>
-        [Test]
-        public static void TestIntersectsSequenceVariations()
-        {
-            Protein protein = new Protein("MACDEFGHIK", "test");
-            PeptideWithSetModifications pepe = new PeptideWithSetModifications(protein, new DigestionParams(), 2, 10, CleavageSpecificity.Unknown, "", 0, new Dictionary<int, Modification>(), 0);
-
-            // The weird thing here is that IntersectsWithVariation takes in applied variations,
-            // so these are constructed as if already applied
-            SequenceVariation sv1Before = new SequenceVariation(1, 1, "A", "M", ""); // before peptide (not identified)
-            SequenceVariation sv2Synonymous = new SequenceVariation(2, 2, "A", "A", ""); // no change (intersects because peptide crosses entire variant but is not truly "identified")
-            SequenceVariation sv4MissenseBeginning = new SequenceVariation(2, 2, "V", "A", ""); // missense at beginning
-            SequenceVariation sv5InsertionAtEnd = new SequenceVariation(7, 9, "GHI", "GHIK", ""); // insertion or stop loss
-            SequenceVariation sv6Deletion = new SequenceVariation(2, 3, "AC", "A", ""); // deletion
-            SequenceVariation sv66Truncation = new SequenceVariation(10, 20, "KAAAAAAAAAA", "K", ""); // truncation or stop gain (identified because peptide crosses entire variant)
-            SequenceVariation sv7MNP = new SequenceVariation(2, 3, "AA", "AC", ""); // mnp
-            SequenceVariation sv77MNP = new SequenceVariation(2, 3, "AC", "AC", ""); // synonymous mnp (identified because peptide crosses entire variant)
-            SequenceVariation sv9MissenseInRange = new SequenceVariation(3, 3, "C", "V", ""); // missense in range
-            SequenceVariation sv10MissenseRangeEdge = new SequenceVariation(10, 10, "K", "R", ""); // missense at end
-            SequenceVariation sv11After = new SequenceVariation(11, 11, "L", "V", ""); // after peptide (not identified)
-
-            Assert.IsFalse(pepe.IntersectsAndIdentifiesVariation(sv1Before).intersects);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv2Synonymous).intersects);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv4MissenseBeginning).intersects);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv5InsertionAtEnd).intersects);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv6Deletion).intersects);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv66Truncation).intersects);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv7MNP).intersects);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv77MNP).intersects);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv9MissenseInRange).intersects);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv10MissenseRangeEdge).intersects);
-            Assert.IsFalse(pepe.IntersectsAndIdentifiesVariation(sv11After).intersects);
-
-            Assert.IsFalse(pepe.IntersectsAndIdentifiesVariation(sv1Before).identifies);
-            Assert.IsFalse(pepe.IntersectsAndIdentifiesVariation(sv2Synonymous).identifies);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv4MissenseBeginning).identifies);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv5InsertionAtEnd).identifies);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv6Deletion).identifies);
-            Assert.IsFalse(pepe.IntersectsAndIdentifiesVariation(sv66Truncation).identifies);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv7MNP).identifies);
-            Assert.IsFalse(pepe.IntersectsAndIdentifiesVariation(sv77MNP).identifies);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv9MissenseInRange).identifies);
-            Assert.IsTrue(pepe.IntersectsAndIdentifiesVariation(sv10MissenseRangeEdge).identifies);
-            Assert.IsFalse(pepe.IntersectsAndIdentifiesVariation(sv11After).identifies);
-
-            PeptideWithSetModifications pepe2 = new PeptideWithSetModifications(protein, new DigestionParams(), 2, 9, CleavageSpecificity.Unknown, "", 0, new Dictionary<int, Modification>(), 0);
-            Assert.IsTrue(pepe2.IntersectsAndIdentifiesVariation(sv5InsertionAtEnd).intersects); // this only intersects GHI, which is the same in GHI -> GHIK
-            Assert.IsFalse(pepe2.IntersectsAndIdentifiesVariation(sv5InsertionAtEnd).identifies);
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests detection of variant-containing peptides.
-        /// The IsVariantPeptide() method must correctly distinguish peptides that
-        /// span applied sequence variations from those that don't. Essential for
-        /// filtering and reporting variant peptides.
-        /// </summary>
-        [Test]
-        public static void TestIsVariantPeptide()
-        {
-            Protein protein = new Protein("MPEPTIDENEWPEPTIDE", "protein0", appliedSequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 4, "P", "V", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) });
-
-            PeptideWithSetModifications pepe = new PeptideWithSetModifications(protein, new DigestionParams(), 1, 8, CleavageSpecificity.Unknown, "", 0, new Dictionary<int, Modification>(), 0);
-            PeptideWithSetModifications notPepe = new PeptideWithSetModifications(protein, new DigestionParams(), 9, 18, CleavageSpecificity.Unknown, "", 0, new Dictionary<int, Modification>(), 0);
-
-            Assert.IsTrue(pepe.IsVariantPeptide());
-            Assert.IsFalse(notPepe.IsVariantPeptide());
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests sequence variant string generation for reporting.
-        /// The SequenceVariantString method must correctly format variant annotations
-        /// including N-terminal mods, middle mods, and truncated variants for output.
-        /// Essential for human-readable variant peptide reporting.
-        /// </summary>
-        [Test]
-        public static void TestSeqVarString()
-        {
-            Protein protein = new Protein("MACDEFGHIK", "test");
-
-            // mod on N-terminus
-            PeptideWithSetModifications pepe = new PeptideWithSetModifications(protein, new DigestionParams(), 1, 10, CleavageSpecificity.Unknown, "", 0, new Dictionary<int, Modification> { { 1, new Modification("mod on M", "mod", "mod", "mod") } }, 0);
-            SequenceVariation sv1Before = new SequenceVariation(1, 1, "A", "M", ""); // n-terminal mod goes before the sequence
-            Assert.AreEqual("A1[mod:mod on M]M", pepe.SequenceVariantString(sv1Before, true));
-
-            // mod in middle
-            PeptideWithSetModifications pepe2 = new PeptideWithSetModifications(protein, new DigestionParams(), 2, 10, CleavageSpecificity.Unknown, "", 0, new Dictionary<int, Modification> { { 2, new Modification("mod on A", "mod", "mod", "mod") } }, 0);
-            SequenceVariation sv4MissenseBeginning = new SequenceVariation(2, 2, "V", "A", ""); // missense at beginning
-            Assert.AreEqual("V2A[mod:mod on A]", pepe2.SequenceVariantString(sv4MissenseBeginning, true));
-
-            // truncated seqvar doesn't truncate in string report (using applied variation correctly)
-            PeptideWithSetModifications pepe3 = new PeptideWithSetModifications(protein, new DigestionParams(), 2, 9, CleavageSpecificity.Unknown, "", 0, new Dictionary<int, Modification>(), 0);
-            SequenceVariation svvvv = new SequenceVariation(7, 10, "GHM", "GHIK", ""); // insertion
-            Assert.AreEqual("GHM7GHIK", pepe3.SequenceVariantString(svvvv, true));
-
-            Protein protein2 = new Protein("WACDEFGHIK", "test");
-
-            //variant starts at protein start but peptide does not
-            PeptideWithSetModifications pepe4 = new PeptideWithSetModifications(protein2, new DigestionParams(), 4, 8, CleavageSpecificity.Unknown, "", 0, new Dictionary<int, Modification>(), 0);
-            SequenceVariation variant = new SequenceVariation(1, 10, "MABCDEFGHIJKLMNOP", "WACDEFGHIK", ""); // frameshift
-            Assert.AreEqual("MABCDEFGHIJKLMNOP1WACDEFGHIK", pepe4.SequenceVariantString(variant, true));
-        }
-
-        /// <summary>
-        /// CRITICAL: Comprehensive test for variant identification and string methods.
-        /// Tests all variant types (SAV, MNV, insertion, deletion, stop-gain, stop-loss)
-        /// for correct intersection detection, identification logic, and string formatting.
-        /// Essential for proteogenomic variant peptide analysis.
-        /// </summary>
-        [Test]
-        public static void TestIdentifyandStringMethods()
-        {
-            ModificationMotif.TryGetMotif("V", out ModificationMotif motifV);
-            Modification mv = new Modification("mod", null, "type", null, motifV, "Anywhere.", null, 42.01, new Dictionary<string, IList<string>>(), null, null, null, null, null);
-            ModificationMotif.TryGetMotif("P", out ModificationMotif motifP);
-            Modification mp = new Modification("mod", null, "type", null, motifP, "Anywhere.", null, 42.01, new Dictionary<string, IList<string>>(), null, null, null, null, null);
-            Dictionary<int, Modification> modV = new Dictionary<int, Modification>();
-            modV.Add(4, mv);
-            Dictionary<int, Modification> modP = new Dictionary<int, Modification>();
-            modP.Add(5, mp);
-
-            Dictionary<int, List<Modification>> proteinPMods = new Dictionary<int, List<Modification>>();
-            proteinPMods.Add(4, new List<Modification>() { mp });
-
-            List<Protein> proteins = new List<Protein>
-            {
-                new Protein("MPEPTIDE", "protein0", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 4, "P", "V", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }),
-                new Protein("MPEPTIDE", "protein1", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 5, "PT", "KT", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }),
-                new Protein("MPEPTIDE", "protein2", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 4, "P", "PPP", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }),
-                new Protein("MPEPPPTIDE", "protein3", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 6, "PPP", "P", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }),
-                new Protein("MPEPKPKTIDE", "protein4", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 7, "PKPK", "PK", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }),
-                new Protein("MPEPTAIDE", "protein5",sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 6, "PTA", "KT", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }),
-                new Protein("MPEKKAIDE", "protein6", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 6, "KKA", "K", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }),
-                new Protein("MPEPTIDE", "protein7", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 4, "P", "V", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", new Dictionary<int, List<Modification>> {{ 4, new[] { mv }.ToList() } }) }),
-                new Protein("MPEPTIDE", "protein8",sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 4, "P", "PPP", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", new Dictionary<int, List<Modification>> {{ 5, new[] { mp }.ToList() } }) }),
-                new Protein("MPEPTIDEPEPTIDE", "protein9", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 15, "PTIDEPEPTIDE", "PPP", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }),
-                new Protein("MPEPTIDE", "protein10", oneBasedModifications: proteinPMods ,sequenceVariations: new List<SequenceVariation> { new SequenceVariation(4, 4, "P", "V", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }),
-                new Protein("MPEPTIDE", "protein11", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(5, 5, "T", "*", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }), //stop-gain (can identify)
-                new Protein("MPEKTIDE", "protein12", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(5, 5, "T", "*", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }), //stop-gain (can't identify)
-                new Protein("MPEPTIPEPEPTIPE", "protein13", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(7, 7, "P", "D", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }),
-                new Protein("MPEPTIDE", "protein14", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(8, 9, "E", "EK", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }), //peptide becomes longer, and cleavage site is created but cannot be identified
-                new Protein("MPEPTIDE", "protein15", sequenceVariations: new List<SequenceVariation> { new SequenceVariation(9, 13, "*", "KMPEP", @"1\t50000000\t.\tA\tG\t.\tPASS\tANN=G||||||||||||||||\tGT:AD:DP\t1/1:30,30:30", null) }), // stop loss at end of original protein that cannot be identified
-            };
-
-            DigestionParams dp = new DigestionParams(minPeptideLength: 2);
-            DigestionParams dp2 = new DigestionParams(protease: "Asp-N", minPeptideLength: 2);
-            DigestionParams dp3 = new DigestionParams(protease: "Lys-N", minPeptideLength: 2);
-
-            var protein0_variant = proteins.ElementAt(0).GetVariantBioPolymers().ElementAt(0);
-            var protein1_variant = proteins.ElementAt(1).GetVariantBioPolymers().ElementAt(0);
-            var protein2_variant = proteins.ElementAt(2).GetVariantBioPolymers().ElementAt(0);
-            var protein3_variant = proteins.ElementAt(3).GetVariantBioPolymers().ElementAt(0);
-            var protein4_variant = proteins.ElementAt(4).GetVariantBioPolymers().ElementAt(0);
-            var protein5_variant = proteins.ElementAt(5).GetVariantBioPolymers().ElementAt(0);
-            var protein6_variant = proteins.ElementAt(6).GetVariantBioPolymers().ElementAt(0);
-            var protein7_variant = proteins.ElementAt(7).GetVariantBioPolymers().ElementAt(0);
-            var protein8_variant = proteins.ElementAt(8).GetVariantBioPolymers().ElementAt(0);
-            var protein9_variant = proteins.ElementAt(9).GetVariantBioPolymers().ElementAt(0);
-            var protein10_variant = proteins.ElementAt(10).GetVariantBioPolymers().ElementAt(0);
-            var protein11_variant = proteins.ElementAt(11).GetVariantBioPolymers().ElementAt(0);
-            var protein12_variant = proteins.ElementAt(12).GetVariantBioPolymers().ElementAt(0);
-            var protein13_variant = proteins.ElementAt(13).GetVariantBioPolymers().ElementAt(0);
-            var protein14_variant = proteins.ElementAt(14).GetVariantBioPolymers().ElementAt(0);
-            var protein15_variant = proteins.ElementAt(15).GetVariantBioPolymers().ElementAt(0);
-
-            List<Modification> digestMods = new List<Modification>();
-
-            var protein0_peptide = protein0_variant.Digest(dp, digestMods, digestMods).ElementAt(0);
-            var protein0_peptide2 = protein0_variant.Digest(dp2, digestMods, digestMods).ElementAt(0);
-            var protein1_peptide = protein1_variant.Digest(dp, digestMods, digestMods).ElementAt(2);
-            var protein2_peptide = protein2_variant.Digest(dp, digestMods, digestMods).ElementAt(0);
-            var protein3_peptide = protein3_variant.Digest(dp, digestMods, digestMods).ElementAt(0);
-            var protein4_peptide = protein4_variant.Digest(dp, digestMods, digestMods).ElementAt(2);
-            var protein5_peptide = protein5_variant.Digest(dp, digestMods, digestMods).ElementAt(2);
-            var protein6_peptide = protein6_variant.Digest(dp, digestMods, digestMods).ElementAt(2);
-            var protein7_peptide = protein7_variant.Digest(dp, digestMods, digestMods).ElementAt(1);
-            var protein8_peptide = protein8_variant.Digest(dp, digestMods, digestMods).ElementAt(1);
-            var protein9_peptide = protein9_variant.Digest(dp, digestMods, digestMods).ElementAt(0);
-            var protein10_peptide = protein10_variant.Digest(dp, digestMods, digestMods).ElementAt(0);
-            var protein11_peptide = protein11_variant.Digest(dp2, digestMods, digestMods).ElementAt(0);
-            var protein11_peptide2 = protein11_variant.Digest(dp, digestMods, digestMods).ElementAt(0);
-            var protein12_peptide = protein12_variant.Digest(dp, digestMods, digestMods).ElementAt(0);
-            var protein13_peptide = protein13_variant.Digest(dp2, digestMods, digestMods).ElementAt(0);
-            var protein14_peptide = protein14_variant.Digest(dp3, digestMods, digestMods).ElementAt(0);
-            var protein15_peptide = protein15_variant.Digest(dp3, digestMods, digestMods).ElementAt(0);
-
-            Assert.AreEqual((true, true), protein0_peptide.IntersectsAndIdentifiesVariation(protein0_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((true, true), protein0_peptide2.IntersectsAndIdentifiesVariation(protein0_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((true, true), protein1_peptide.IntersectsAndIdentifiesVariation(protein1_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((true, true), protein2_peptide.IntersectsAndIdentifiesVariation(protein2_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((true, true), protein3_peptide.IntersectsAndIdentifiesVariation(protein3_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((false, false), protein4_peptide.IntersectsAndIdentifiesVariation(protein4_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((true, true), protein5_peptide.IntersectsAndIdentifiesVariation(protein5_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((false, true), protein6_peptide.IntersectsAndIdentifiesVariation(protein6_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((true, true), protein7_peptide.IntersectsAndIdentifiesVariation(protein7_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((true, true), protein8_peptide.IntersectsAndIdentifiesVariation(protein8_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((true, true), protein9_peptide.IntersectsAndIdentifiesVariation(protein9_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((true, true), protein10_peptide.IntersectsAndIdentifiesVariation(protein10_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((false, true), protein11_peptide.IntersectsAndIdentifiesVariation(protein11_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((false, true), protein11_peptide2.IntersectsAndIdentifiesVariation(protein11_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((false, false), protein12_peptide.IntersectsAndIdentifiesVariation(protein12_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((false, true), protein13_peptide.IntersectsAndIdentifiesVariation(protein13_variant.AppliedSequenceVariations.ElementAt(0)));
-            Assert.AreEqual((true, false), protein14_peptide.IntersectsAndIdentifiesVariation(protein14_variant.AppliedSequenceVariations.ElementAt(0)));// the peptide crosses the variant but the newly genrated cleavage site makes the same peptide as without the variant
-            Assert.AreEqual((false, false), protein15_peptide.IntersectsAndIdentifiesVariation(protein15_variant.AppliedSequenceVariations.ElementAt(0)));// the peptide does not cross the variant, and the stop loss adds addition amino acids, but it creates the same peptide as without the variant
-
-            Assert.AreEqual("P4V", protein0_peptide.SequenceVariantString(protein0_variant.AppliedSequenceVariations.ElementAt(0), true));
-            Assert.AreEqual("P4V", protein0_peptide2.SequenceVariantString(protein0_variant.AppliedSequenceVariations.ElementAt(0), true));
-            Assert.AreEqual("PT4KT", protein1_peptide.SequenceVariantString(protein1_variant.AppliedSequenceVariations.ElementAt(0), true));
-            Assert.AreEqual("P4PPP", protein2_peptide.SequenceVariantString(protein2_variant.AppliedSequenceVariations.ElementAt(0), true));
-            Assert.AreEqual("PPP4P", protein3_peptide.SequenceVariantString(protein3_variant.AppliedSequenceVariations.ElementAt(0), true));
-            Assert.AreEqual("PTA4KT", protein5_peptide.SequenceVariantString(protein5_variant.AppliedSequenceVariations.ElementAt(0), true));
-            Assert.AreEqual("KKA4K", protein6_peptide.SequenceVariantString(protein6_variant.AppliedSequenceVariations.ElementAt(0), false));
-            Assert.AreEqual("P4V[type:mod on V]", protein7_peptide.SequenceVariantString(protein7_variant.AppliedSequenceVariations.ElementAt(0), true));
-            Assert.AreEqual("P4PP[type:mod on P]P", protein8_peptide.SequenceVariantString(protein8_variant.AppliedSequenceVariations.ElementAt(0), true));
-            Assert.AreEqual("PTIDEPEPTIDE4PPP", protein9_peptide.SequenceVariantString(protein9_variant.AppliedSequenceVariations.ElementAt(0), true));
-            Assert.AreEqual("P4V", protein10_peptide.SequenceVariantString(protein10_variant.AppliedSequenceVariations.ElementAt(0), true));
-            Assert.AreEqual("T5*", protein11_peptide.SequenceVariantString(protein11_variant.AppliedSequenceVariations.ElementAt(0), false));
-            Assert.AreEqual("T5*", protein11_peptide2.SequenceVariantString(protein11_variant.AppliedSequenceVariations.ElementAt(0), false));
-            Assert.AreEqual("P7D", protein13_peptide.SequenceVariantString(protein13_variant.AppliedSequenceVariations.ElementAt(0), false));
-        }
-
-        /// <summary>
-        /// DEFENSIVE: Tests error handling for malformed peptide sequences.
-        /// Validates that ambiguous sequences, bad modifications, and nonexistent mods
-        /// throw MzLibException rather than crashing or returning invalid peptides.
-        /// </summary>
-        [Test]
-        public static void BreakDeserializationMethod()
-        {
-            Assert.Throws<MzLibUtil.MzLibException>(() => new PeptideWithSetModifications("|", new Dictionary<string, Modification>())); // ambiguous
-            Assert.Throws<MzLibUtil.MzLibException>(() => new PeptideWithSetModifications("[]", new Dictionary<string, Modification>())); // bad mod
-            Assert.Throws<MzLibUtil.MzLibException>(() => new PeptideWithSetModifications("A[:mod]", new Dictionary<string, Modification>())); // nonexistent mod
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests reverse decoy peptide generation from target peptides.
-        /// Validates correct sequence reversal, modification position remapping,
-        /// cleavage site preservation, and handling of palindromic sequences.
-        /// Essential for target-decoy FDR estimation in all database searches.
-        /// </summary>
-        [Test]
-        public static void TestReverseDecoyFromTarget()
-        {
-            ModificationMotif.TryGetMotif("P", out ModificationMotif motif_p);
-            Modification phosphorylation = new Modification(_originalId: "phospho", _modificationType: "CommonBiological", _target: motif_p, _locationRestriction: "Anywhere.", _chemicalFormula: ChemicalFormula.ParseFormula("H1O3P1"));
-
-            ModificationMotif.TryGetMotif("K", out ModificationMotif motif_k);
-            Modification acetylation = new Modification(_originalId: "acetyl", _modificationType: "CommonBiological", _target: motif_k, _locationRestriction: "Anywhere.", _chemicalFormula: ChemicalFormula.ParseFormula("H1O3P1"));
-
-            Modification nTermAcet = new Modification(_originalId: "n-acetyl", _modificationType: "CommonBiological", _target: null, _locationRestriction: "N-terminal.", _chemicalFormula: ChemicalFormula.ParseFormula("H"));
-
-            Dictionary<int, Modification> allmodsoneisnterminus = new Dictionary<int, Modification> { { 1, nTermAcet }, { 6, phosphorylation }, { 9, acetylation } };
-
-            PeptideWithSetModifications p = new PeptideWithSetModifications(new Protein("PEPTIDEK", "ACCESSIION"), new DigestionParams(), 1, 8, CleavageSpecificity.Full, null, 0, allmodsoneisnterminus, 0, null);
-
-            int[] newAminoAcidPositions = new int["PEPTIDEK".Length];
-            PeptideWithSetModifications reverse = p.GetReverseDecoyFromTarget(newAminoAcidPositions);
-
-            string targetSequence = p.FullSequence;
-            string decoySequence = reverse.FullSequence; 
-            Assert.AreEqual(reverse.PairedTargetDecoySequence, targetSequence);
-            Assert.AreEqual(p.PairedTargetDecoySequence, decoySequence);
-            Assert.AreEqual("EDITPEPK", reverse.BaseSequence);
-            Assert.AreEqual(new int[] { 6, 5, 4, 3, 2, 1, 0, 7 }, newAminoAcidPositions);
-            Assert.IsTrue(reverse.Protein.IsDecoy);
-            Assert.AreEqual(p.Protein.BaseSequence.Length, reverse.Protein.BaseSequence.Length);//we replaced the original with the new so the protein should have the same length
-            Assert.AreEqual(reverse.PeptideDescription, p.FullSequence);
-
-            List<Product> decoyProducts = new List<Product>();
-            reverse.Fragment(MassSpectrometry.DissociationType.HCD, FragmentationTerminus.Both, decoyProducts);
-
-            Assert.AreEqual(14, decoyProducts.Count);
-
-            //  Arg-C -- Cleave after R
-            newAminoAcidPositions = new int["RPEPTIREK".Length];
-            PeptideWithSetModifications p_argC = new PeptideWithSetModifications(new Protein("RPEPTIREK", "DECOY_ARGC"), new DigestionParams(protease: "Arg-C"), 1, 9, CleavageSpecificity.Full, null, 0, new Dictionary<int, Modification>(), 0, null);
-            PeptideWithSetModifications p_argC_reverse = p_argC.GetReverseDecoyFromTarget(newAminoAcidPositions);
-            Assert.AreEqual("RKEITPREP", p_argC_reverse.BaseSequence);
-            Assert.AreEqual(new int[] { 0, 8, 7, 5, 4, 3, 6, 2, 1 }, newAminoAcidPositions);
-            Assert.AreEqual(p_argC_reverse.PeptideDescription, p_argC.FullSequence);
-
-            //  Asp-N -- Cleave before D
-            newAminoAcidPositions = new int["DPEPTIDEK".Length];
-            PeptideWithSetModifications p_aspN = new PeptideWithSetModifications(new Protein("DPEPTIDEK", "DECOY_ASPN"), new DigestionParams(protease: "Asp-N"), 1, 9, CleavageSpecificity.Full, null, 0, new Dictionary<int, Modification>(), 0, null);
-            PeptideWithSetModifications p_aspN_reverse = p_aspN.GetReverseDecoyFromTarget(newAminoAcidPositions);
-            Assert.AreEqual("DKEITPDEP", p_aspN_reverse.BaseSequence);
-            Assert.AreEqual(p_aspN.FullSequence, p_aspN_reverse.PeptideDescription);
-
-            //  chymotrypsin|P
-            newAminoAcidPositions = new int["FKFPRWAWPSYGYPG".Length];
-            PeptideWithSetModifications p_chymoP = new PeptideWithSetModifications(new Protein("FKFPRWAWPSYGYPG", "DECOY_CHYMOP"), new DigestionParams(protease: "chymotrypsin|P", maxMissedCleavages: 10), 1, 15, CleavageSpecificity.Full, null, 0, new Dictionary<int, Modification>(), 0, null);
-            PeptideWithSetModifications p_chymoP_reverse = p_chymoP.GetReverseDecoyFromTarget(newAminoAcidPositions);
-            Assert.AreEqual("FGPYGWSPWAYRPFK", p_chymoP_reverse.BaseSequence);
-            Assert.AreEqual(p_chymoP.FullSequence, p_chymoP_reverse.PeptideDescription);
-
-            //  CNBr cleave after M
-            newAminoAcidPositions = new int["MPEPTIMEK".Length];
-            PeptideWithSetModifications p_cnbr = new PeptideWithSetModifications(new Protein("MPEPTIMEK", "DECOY_CNBR"), new DigestionParams(protease: "CNBr"), 1, 9, CleavageSpecificity.Full, null, 0, new Dictionary<int, Modification>(), 0, null);
-            PeptideWithSetModifications p_cnbr_reverse = p_cnbr.GetReverseDecoyFromTarget(newAminoAcidPositions);
-            Assert.AreEqual("MKEITPMEP", p_cnbr_reverse.BaseSequence);
-            Assert.AreEqual(p_cnbr.FullSequence, p_cnbr_reverse.PeptideDescription);
-
-            //  elastase cleave after A, V, S, G, L, I,
-            newAminoAcidPositions = new int["KAYVPSRGHLDIN".Length];
-            PeptideWithSetModifications p_elastase = new PeptideWithSetModifications(new Protein("KAYVPSRGHLDIN", "DECOY_ELASTASE"), new DigestionParams(protease: "elastase|P"), 1, 13, CleavageSpecificity.Semi, null, 0, new Dictionary<int, Modification>(), 0, null);
-            PeptideWithSetModifications p_elastase_reverse = p_elastase.GetReverseDecoyFromTarget(newAminoAcidPositions);
-            Assert.AreEqual("NADHRSPGVLYIK", p_elastase_reverse.BaseSequence);
-
-            //  top-down
-            newAminoAcidPositions = new int["RPEPTIREK".Length];
-            PeptideWithSetModifications p_topdown = new PeptideWithSetModifications(new Protein("RPEPTIREK", "DECOY_TD"), new DigestionParams(protease: "top-down"), 1, 9, CleavageSpecificity.Full, null, 0, new Dictionary<int, Modification>(), 0, null);
-            PeptideWithSetModifications p_topdown_reverse = p_topdown.GetReverseDecoyFromTarget(newAminoAcidPositions);
-            Assert.AreEqual("KERITPEPR", p_topdown_reverse.BaseSequence);
-
-            //  Arg-C -- Cleave after R
-            newAminoAcidPositions = new int["PEPTGPYGPYIDE".Length];
-            PeptideWithSetModifications p_coll = new PeptideWithSetModifications(new Protein("PEPTGPYGPYIDE", "DECOY_COL"), new DigestionParams(protease: "collagenase"), 1, 13, CleavageSpecificity.Full, null, 0, new Dictionary<int, Modification>(), 0, null);
-            PeptideWithSetModifications p_coll_reverse = p_coll.GetReverseDecoyFromTarget(newAminoAcidPositions);
-            Assert.AreEqual("EDITGPYGPYPEP", p_coll_reverse.BaseSequence);
-
-            // Tyrpsin -- Reverse Decoy is identical so get mirror
-            Dictionary<int, Modification> VTIRTVR_modsDictionary = new Dictionary<int, Modification> { { 1, nTermAcet }, { 3, phosphorylation } };
-            newAminoAcidPositions = new int["VTIRTVR".Length];
-            PeptideWithSetModifications p_tryp = new PeptideWithSetModifications(new Protein("VTIRTVR", "DECOY_TRYP"), new DigestionParams(protease: "trypsin"), 1, 7, CleavageSpecificity.Full, null, 0, VTIRTVR_modsDictionary, 0, null);
-            PeptideWithSetModifications p_tryp_reverse = p_tryp.GetReverseDecoyFromTarget(newAminoAcidPositions);
-            // Hash code corresponding to the target sequence, should be PairedTargetDecoyHash for reverse
-            string mirrorTarget = p_tryp.FullSequence;
-            // Hash code corresponding to the decoy sequence, should be PairedTargetDecoyHash for target
-            string mirrorDecoy = p_tryp_reverse.FullSequence;
-            Assert.AreEqual(mirrorTarget, p_tryp_reverse.PairedTargetDecoySequence);
-            Assert.AreEqual(mirrorDecoy, p_tryp.PairedTargetDecoySequence);
-            Assert.AreEqual("RVTRITV", p_tryp_reverse.BaseSequence);
-            Assert.AreEqual(new int[] { 6, 5, 4, 3, 2, 1, 0 }, newAminoAcidPositions);
-            Assert.IsTrue(p_tryp_reverse.AllModsOneIsNterminus.ContainsKey(1));//n-term acetyl
-            Assert.IsTrue(p_tryp_reverse.AllModsOneIsNterminus.ContainsKey(7));//moved T-phospho from 3 to 7
-            Assert.IsTrue(p_tryp_reverse.Protein.IsDecoy);
-            Assert.AreEqual(p_tryp.FullSequence, p_tryp_reverse.PeptideDescription);
-
-        }
-        /// <summary>
-        /// CRITICAL: Tests scrambled decoy peptide generation from target peptides.
-        /// Scrambled decoys are an alternative to reversed decoys that may provide
-        /// better FDR estimation for certain peptide populations. Tests modification
-        /// position mapping and cleavage site preservation during scrambling.
-        /// </summary>
-        [Test]
-        public static void TestScrambledDecoyFromTarget()
-        {
-            ModificationMotif.TryGetMotif("P", out ModificationMotif motif_p);
-            Modification phosphorylation = new Modification(_originalId: "phospho", _modificationType: "CommonBiological", _target: motif_p, _locationRestriction: "Anywhere.", _chemicalFormula: ChemicalFormula.ParseFormula("H1O3P1"));
-
-            ModificationMotif.TryGetMotif("K", out ModificationMotif motif_k);
-            Modification acetylation = new Modification(_originalId: "acetyl", _modificationType: "CommonBiological", _target: motif_k, _locationRestriction: "Anywhere.", _chemicalFormula: ChemicalFormula.ParseFormula("H1O3P1"));
-
-            Modification nTermAcet = new Modification(_originalId: "n-acetyl", _modificationType: "CommonBiological", _target: null, _locationRestriction: "N-terminal.", _chemicalFormula: ChemicalFormula.ParseFormula("H"));
-
-            Dictionary<int, Modification> allmodsoneisnterminus = new Dictionary<int, Modification> { { 1, nTermAcet }, { 5, phosphorylation }, { 9, acetylation } };
-
-            PeptideWithSetModifications p = new PeptideWithSetModifications(new Protein("PEPTIDEK", "ACCESSIION"), new DigestionParams(), 1, 8, CleavageSpecificity.Full, null, 0, allmodsoneisnterminus, 0, null);
-            int[] newAminoAcidPositions = new int["PEPTIDEK".Length];
-            PeptideWithSetModifications testScrambled = p.GetScrambledDecoyFromTarget(newAminoAcidPositions);
-
-            string targetSequence = p.FullSequence;
-            string decoySequence = testScrambled.FullSequence;
-            Assert.AreEqual(testScrambled.PairedTargetDecoySequence, targetSequence);
-            Assert.AreEqual(p.PairedTargetDecoySequence, decoySequence);
-            Assert.AreEqual("IDEETPPK", testScrambled.BaseSequence);
-            Assert.AreEqual(new int[] { 4, 5, 6, 1, 3, 0, 2, 7 }, newAminoAcidPositions);
-            // Check n-term acetyl
-            Assert.AreEqual(p.AllModsOneIsNterminus[1], testScrambled.AllModsOneIsNterminus[1]); 
-            // Check phosphorylated Thr originally at position 5
-            Assert.AreEqual(p.AllModsOneIsNterminus[5], testScrambled.AllModsOneIsNterminus[6]);
-            Assert.IsTrue(testScrambled.Protein.IsDecoy);
-            Assert.AreEqual(p.Protein.BaseSequence.Length, testScrambled.Protein.BaseSequence.Length);
-            // Check that the scrambled PeptideDescription is equivalent to the original peptide's full sequence
-            Assert.AreEqual(testScrambled.PeptideDescription, p.FullSequence);
-
-            // Test for complex cleavage motif patterns
-            Dictionary<int, Modification> complexTrypticMods = new Dictionary<int, Modification> { { 1, nTermAcet }, { 89, acetylation } };
-            string insulinFile = Path.Combine(TestContext.CurrentContext.TestDirectory, "DataFiles", "humanInsulin.fasta");
-            string insulinSeq = ProteinDbLoader.LoadProteinFasta(insulinFile, true, DecoyType.None, false, out var dbError)[0].BaseSequence;
-            PeptideWithSetModifications insulinTryptic = new PeptideWithSetModifications(new Protein(insulinSeq, "COMPLEX_TRYP"), new DigestionParams(), 1, 110, CleavageSpecificity.Full, null, 0, complexTrypticMods, 0, null);
-            int[] insulinScrambledOrder = new int[insulinSeq.Length];
-            PeptideWithSetModifications insulinScrambled = insulinTryptic.GetScrambledDecoyFromTarget(insulinScrambledOrder);
-            // Check that cleavage motif positions are preserved in the more complex case
-            Assert.AreEqual(insulinTryptic.BaseSequence[5], insulinScrambled[5]);
-            Assert.AreEqual(insulinScrambled.BaseSequence[55], insulinTryptic.BaseSequence[55]);          
-            // Check that modified cleavage motif residues are preserved
-            Assert.AreEqual(insulinScrambled.AllModsOneIsNterminus[89], insulinTryptic.AllModsOneIsNterminus[89]);
-
-            // Test with Arg-C as the protease
-            newAminoAcidPositions = new int["RPEPTIREAVLKK".Length];
-            PeptideWithSetModifications testArgC = new PeptideWithSetModifications(new Protein("RPEPTIREAVLKK", "DECOY_ARGC"), new DigestionParams(protease: "Arg-C"), 1, 13, CleavageSpecificity.Full, null, 0, new Dictionary<int, Modification>(), 0, null);
-            PeptideWithSetModifications scrambledArgC = testArgC.GetScrambledDecoyFromTarget(newAminoAcidPositions);
-            // Check for preserved cleavage motif positions
-            Assert.AreEqual(testArgC.BaseSequence[6], scrambledArgC.BaseSequence[6]);
-
-            // Tests that peptide is mirrored when the maximum number of scramble attempts is reached
-            newAminoAcidPositions = new int["AVLRRRKKRDEF".Length];
-            PeptideWithSetModifications forceMirror = new PeptideWithSetModifications(new Protein("AVLRRRKKRDEF", "ACCESSIION"), new DigestionParams(), 1, 12, CleavageSpecificity.Full, null, 0, allmodsoneisnterminus, 0, null);
-            PeptideWithSetModifications mirroredTarget = forceMirror.GetScrambledDecoyFromTarget(newAminoAcidPositions);
-            Assert.AreEqual(new int[] { 11, 10, 9, 8, 7, 6, 5, 4, 3, 2, 1, 0 }, newAminoAcidPositions);
-        }
-        /// <summary>
-        /// CRITICAL: Tests decoy generation from peptides digested from real XML databases.
-        /// Ensures decoy generation works correctly with database-derived modifications,
-        /// sequence variants, and other annotations. No unchanged peptides should remain
-        /// after decoy generation (except true palindromes).
-        /// </summary>
-        [Test]
-        public static void TestReverseDecoyFromPeptideFromProteinXML()
-        {
-            //Just making sure there are no snafus when creating decoy peptides from an xml,which will have mods in various places, etc.
-            //sequence variants, modifications
-            Dictionary<string, Modification> un = new Dictionary<string, Modification>();
-            var psiModDeserialized = Loaders.LoadPsiMod(Path.Combine(TestContext.CurrentContext.TestDirectory, "PSI-MOD.obo2.xml"));
-            Dictionary<string, int> formalChargesDictionary = Loaders.GetFormalChargesDictionary(psiModDeserialized);
-            List<Modification> UniProtPtms = Loaders.LoadUniprot(Path.Combine(TestContext.CurrentContext.TestDirectory, "ptmlist2.txt"), formalChargesDictionary).ToList();
-            List<Protein> proteins = ProteinDbLoader.LoadProteinXML(Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "cRAP_databaseGPTMD.xml"), true, DecoyType.None, UniProtPtms, false, new string[] { "exclude_me" }, out un);
-
-            List<Modification> fixedMods = new List<Modification>();
-            List<Modification> variableMods = new List<Modification>();
-            ModificationMotif.TryGetMotif("C", out ModificationMotif motif_C);
-            ModificationMotif.TryGetMotif("M", out ModificationMotif motif_M);
-
-            fixedMods.Add(new Modification(_originalId: "resMod_C", _target: motif_C, _locationRestriction: "Anywhere.", _chemicalFormula: ChemicalFormula.ParseFormula("H"), _monoisotopicMass: PeriodicTable.GetElement(1).PrincipalIsotope.AtomicMass));
-            fixedMods.Add(new Modification(_originalId: "resMod_M", _target: motif_C, _locationRestriction: "Anywhere.", _chemicalFormula: ChemicalFormula.ParseFormula("O"), _monoisotopicMass: PeriodicTable.GetElement(8).PrincipalIsotope.AtomicMass));
-
-            int unchangedPeptides = 0;
-            int totalPeptides = 0;
-
-            foreach (Protein p in proteins)
-            {
-                List<PeptideWithSetModifications> targetPeptides = p.Digest(new DigestionParams(), fixedMods, variableMods, null, null).ToList();
-                foreach (PeptideWithSetModifications targetPeptide in targetPeptides)
-                {
-                    totalPeptides++;
-                    int[] newAminoAcidPositions = new int[targetPeptide.BaseSequence.Length];
-                    PeptideWithSetModifications decoyPeptide = targetPeptide.GetReverseDecoyFromTarget(newAminoAcidPositions);
-
-                    if (decoyPeptide.BaseSequence == targetPeptide.BaseSequence)
-                    {
-                        unchangedPeptides++;
-                    }
-                }
-            }
-
-            Assert.AreEqual(0, unchangedPeptides);
-        }
-
-        /// <summary>
-        /// INFORMATIONAL: Counts how many target peptides have matching decoy sequences.
-        /// This metric helps assess decoy database quality - too many matches indicate
-        /// potential FDR estimation issues. Used for database quality validation.
-        /// </summary>
-        [Test]
-        public static void CountTargetsWithMatchingDecoys()
-        {
-            Dictionary<string, Modification> un = new Dictionary<string, Modification>();
-            var psiModDeserialized = Loaders.LoadPsiMod(Path.Combine(TestContext.CurrentContext.TestDirectory, "PSI-MOD.obo2.xml"));
-            Dictionary<string, int> formalChargesDictionary = Loaders.GetFormalChargesDictionary(psiModDeserialized);
-            List<Modification> UniProtPtms = Loaders.LoadUniprot(Path.Combine(TestContext.CurrentContext.TestDirectory, "ptmlist2.txt"), formalChargesDictionary).ToList();
-            List<Protein> proteins = ProteinDbLoader.LoadProteinXML(Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "cRAP_databaseGPTMD.xml"), true, DecoyType.None, UniProtPtms, false, new string[] { "exclude_me" }, out un);
-
-            List<Modification> fixedMods = new List<Modification>();
-            List<Modification> variableMods = new List<Modification>();
-            ModificationMotif.TryGetMotif("C", out ModificationMotif motif_C);
-            ModificationMotif.TryGetMotif("M", out ModificationMotif motif_M);
-
-            fixedMods.Add(new Modification(_originalId: "resMod_C", _target: motif_C, _locationRestriction: "Anywhere.", _chemicalFormula: ChemicalFormula.ParseFormula("H"), _monoisotopicMass: PeriodicTable.GetElement(1).PrincipalIsotope.AtomicMass));
-            fixedMods.Add(new Modification(_originalId: "resMod_M", _target: motif_C, _locationRestriction: "Anywhere.", _chemicalFormula: ChemicalFormula.ParseFormula("O"), _monoisotopicMass: PeriodicTable.GetElement(8).PrincipalIsotope.AtomicMass));
-
-            Dictionary<string, int> targets = new Dictionary<string, int>();
-
-            foreach (Protein p in proteins)
-            {
-                List<PeptideWithSetModifications> targetPeptides = p.Digest(new DigestionParams(), fixedMods, variableMods, null, null).ToList();
-
-                foreach (PeptideWithSetModifications targetPeptide in targetPeptides)
-                {
-                    if (targets.ContainsKey(targetPeptide.BaseSequence))
-                    {
-                        targets[targetPeptide.BaseSequence]++;
-                    }
-                    else
-                    {
-                        targets.Add(targetPeptide.BaseSequence, 1);
-                    }
-                }
-            }
-
-            int matchingDecoys = 0;
-            foreach (Protein p in proteins)
-            {
-                List<PeptideWithSetModifications> targetPeptides = p.Digest(new DigestionParams(), fixedMods, variableMods, null, null).ToList();
-
-                foreach (PeptideWithSetModifications target in targetPeptides)
-                {
-                    int[] newAminoAcidPositions = new int[target.BaseSequence.Length];
-                    string decoySequence = target.GetReverseDecoyFromTarget(newAminoAcidPositions).BaseSequence;
-
-                    if (targets.ContainsKey(decoySequence))
-                    {
-                        matchingDecoys++;
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests that top-down digestion returns expected truncation products.
-        /// For insulin, expects 68 truncation products. Essential for top-down and
-        /// middle-down proteomics workflows that search for proteoforms.
-        /// </summary>
-        [Test]
-        public static void TestPeptideWithSetModsReturnsTruncationsInTopDown()
-        {
-            string xmlDatabase = Path.Combine(TestContext.CurrentContext.TestDirectory, "DataFiles", "humanInsulin.xml");
-
-            Protein insulin = ProteinDbLoader.LoadProteinXML(xmlDatabase, true,
-                DecoyType.None, null, false, null, out var unknownModifications, addTruncations: true)[0];
-
-            Protease protease = new Protease("top-down", CleavageSpecificity.None, "", "", new List<DigestionMotif>(), null);
-            List<PeptideWithSetModifications> insulinTruncations = insulin.Digest(new DigestionParams(protease: protease.Name), new List<Modification>(), new List<Modification>(), topDownTruncationSearch: true).ToList();
-            Assert.AreEqual(68, insulinTruncations.Count);
-        }
-        /// <summary>
-        /// CRITICAL: Regression test for top-down truncation product generation.
-        /// Generates a comprehensive table of all truncation peptides for target/decoy
-        /// insulin and compares against expected baseline. Ensures truncation logic
-        /// remains stable across code changes. Essential for top-down proteomics.
-        /// </summary>
-        [Test]
-        public static void TestTopDownTruncationTableMatchesExpected()
-        {
-            // PURPOSE
-            // Generate a comprehensive TSV of all top-down truncation peptides for target and decoy insulin entries,
-            // then compare the result against the stored expected file for regression checking.
-            //
-            // OUTPUT COLUMNS
-            // - Sequence: peptide/proteoform base sequence
-            // - Type: "Target" or "Decoy" based on parent protein
-            // - Begin: one-based start residue within the parent protein
-            // - End: one-based end residue within the parent protein
-            // - RetainedMethionine: TRUE when the peptide includes the protein’s N-terminal Met (Begin == 1 and Parent.BaseSequence[0] == 'M'), else FALSE
-
-            // Arrange: load insulin with reverse decoys and truncations enabled
-            string xmlDatabase = Path.Combine(TestContext.CurrentContext.TestDirectory, "DataFiles", "humanInsulin.xml");
-            List<Protein> insulinProteins = ProteinDbLoader.LoadProteinXML(
-                xmlDatabase,
-                generateTargets: true,
-                decoyType: DecoyType.Reverse,
-                allKnownModifications: null,
-                isContaminant: false,
-                modTypesToExclude: null,
-                unknownModifications: out var unknownModifications,
-                addTruncations: true);
-
-            Assert.That(insulinProteins.Any(p => !p.IsDecoy), "Expected at least one target protein");
-            Assert.That(insulinProteins.Any(p => p.IsDecoy), "Expected at least one decoy protein");
-            Assert.That(unknownModifications == null || unknownModifications.Count == 0, "No unknown modifications expected from insulin XML");
-
-            // Digest: enumerate truncations for a representative target/decoy pair (parity sanity)
-            static string Reverse(string s) => new string(s.Reverse().ToArray());
-            var target = insulinProteins.First(p => !p.IsDecoy);
-            string expectedDecoySeq = target.BaseSequence.Length > 0 && target.BaseSequence[0] == 'M'
-                ? "M" + Reverse(target.BaseSequence.Substring(1))
-                : Reverse(target.BaseSequence);
-            var decoy = insulinProteins.FirstOrDefault(p => p.IsDecoy && p.BaseSequence == expectedDecoySeq)
-                        ?? insulinProteins.First(p => p.IsDecoy && p.Length == target.Length);
-
-            Assert.IsFalse(string.IsNullOrWhiteSpace(target.Accession));
-            Assert.IsTrue(decoy.Accession.StartsWith("DECOY_"));
-            Assert.AreEqual(target.Length, decoy.Length);
-            Assert.AreNotEqual(target.BaseSequence, decoy.BaseSequence);
-            Assert.AreEqual(expectedDecoySeq, decoy.BaseSequence, "Decoy must follow 'retain M, reverse remainder' rule");
-
-            var dp = new DigestionParams(protease: "top-down");
-            List<PeptideWithSetModifications> targetTruncs = target
-                .Digest(dp, new List<Modification>(), new List<Modification>(), topDownTruncationSearch: true)
-                .Cast<PeptideWithSetModifications>().ToList();
-            List<PeptideWithSetModifications> decoyTruncs = decoy
-                .Digest(dp, new List<Modification>(), new List<Modification>(), topDownTruncationSearch: true)
-                .Cast<PeptideWithSetModifications>().ToList();
-
-            // Parity and sanity checks for the selected pair
-            Assert.AreEqual(68, targetTruncs.Count, "Target should yield 68 truncation products in top-down mode");
-            Assert.AreEqual(68, decoyTruncs.Count, "Decoy should yield 68 truncation products in top-down mode");
-            Assert.That(targetTruncs.All(p => p.DigestionParams?.DigestionAgent?.Name == "top-down"));
-            Assert.That(decoyTruncs.All(p => p.DigestionParams?.DigestionAgent?.Name == "top-down"));
-            Assert.AreEqual(targetTruncs.Count, targetTruncs.Select(p => p.BaseSequence).Distinct().Count());
-            Assert.AreEqual(decoyTruncs.Count, decoyTruncs.Select(p => p.BaseSequence).Distinct().Count());
-            Assert.IsTrue(targetTruncs.Any(p => p.BaseSequence == target.BaseSequence));
-            Assert.IsTrue(decoyTruncs.Any(p => p.BaseSequence == decoy.BaseSequence));
-
-            // Build the table rows
-            static bool HasRetainedMet(PeptideWithSetModifications p) =>
-                p.OneBasedStartResidueInProtein == 1 &&
-                p.Parent?.BaseSequence?.Length > 0 &&
-                p.Parent.BaseSequence[0] == 'M';
-
-            // We only compare the combined truncations for the chosen target/decoy in this test (68 + 68 rows)
-            var rows = targetTruncs.Concat(decoyTruncs)
-                .Select(pep =>
-                {
-                    bool isDecoy = (pep.Parent as Protein)?.IsDecoy == true;
-                    string type = isDecoy ? "Decoy" : "Target";
-                    string retained = HasRetainedMet(pep) ? "TRUE" : "FALSE"; // normalize to match expected
-                    return string.Join("\t", new[]
-                    {
-                        pep.BaseSequence,
-                        type,
-                        pep.OneBasedStartResidueInProtein.ToString(),
-                        pep.OneBasedEndResidueInProtein.ToString(),
-                        retained
-                    });
-                })
-                // Sort deterministically to avoid platform/iteration-order differences
-                .OrderBy(r => r, StringComparer.Ordinal)
-                .ToList();
-
-            var header = "Sequence\tType\tBegin\tEnd\tRetainedMethionine";
-            var outputLines = new List<string>(capacity: rows.Count + 1) { header };
-            outputLines.AddRange(rows);
-
-            // Persist the generated table for inspection
-            string workPath = Path.Combine(TestContext.CurrentContext.WorkDirectory, "topdown_truncations_table.tsv");
-            File.WriteAllLines(workPath, outputLines);
-            Console.WriteLine($"Generated truncation table: {workPath}");
-
-            // Load expected and compare
-            string expectedPath = Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "truncationsExpected.tsv");
-            Assert.That(File.Exists(expectedPath), $"Expected file not found: {expectedPath}");
-            var expectedAll = File.ReadAllLines(expectedPath)
-                                  .Where(l => l is not null)
-                                  .Select(l => l.TrimEnd('\r', '\n'))
-                                  .ToList();
-
-            Assert.That(expectedAll.Count > 0, "Expected file is empty");
-            string expectedHeader = expectedAll[0];
-            var expectedRows = expectedAll.Skip(1)
-                                          .Where(l => !string.IsNullOrWhiteSpace(l))
-                                          .OrderBy(l => l, StringComparer.Ordinal)
-                                          .ToList();
-
-            // Header check
-            if (!string.Equals(header, expectedHeader, StringComparison.Ordinal))
-            {
-                TestContext.Out.WriteLine($"Header mismatch:");
-                TestContext.Out.WriteLine($"  Expected: {expectedHeader}");
-                TestContext.Out.WriteLine($"  Actual  : {header}");
-            }
-
-            // Multiset comparison for rows (counts of duplicates matter)
-            static Dictionary<string, int> ToCounts(IEnumerable<string> lines)
-                => lines.GroupBy(x => x, StringComparer.Ordinal)
-                        .ToDictionary(g => g.Key, g => g.Count(), StringComparer.Ordinal);
-
-            var expCounts = ToCounts(expectedRows);
-            var gotCounts = ToCounts(rows);
-
-            var missing = new List<string>();   // in expected more times than in actual
-            var extra = new List<string>();     // in actual more times than in expected
-
-            foreach (var kv in expCounts)
-            {
-                gotCounts.TryGetValue(kv.Key, out int got);
-                if (got < kv.Value)
-                {
-                    int deficit = kv.Value - got;
-                    for (int i = 0; i < deficit; i++) missing.Add(kv.Key);
-                }
-            }
-            foreach (var kv in gotCounts)
-            {
-                expCounts.TryGetValue(kv.Key, out int exp);
-                if (kv.Value > exp)
-                {
-                    int surplus = kv.Value - exp;
-                    for (int i = 0; i < surplus; i++) extra.Add(kv.Key);
-                }
-            }
-
-            if (missing.Count == 0 && extra.Count == 0)
-            {
-                TestContext.Out.WriteLine("Top-down truncation table matches expected.");
-                TestContext.Out.WriteLine("Sample (first 5 rows):");
-                foreach (var l in outputLines.Take(6)) TestContext.Out.WriteLine(l);
-            }
-            else
-            {
-                TestContext.Out.WriteLine("Top-down truncation table differs from expected.");
-                TestContext.Out.WriteLine($"Missing rows (expected but not found or under-counted): {missing.Count}");
-                foreach (var l in missing.Take(20)) TestContext.Out.WriteLine($"  MISSING: {l}");
-                if (missing.Count > 20) TestContext.Out.WriteLine($"  ...and {missing.Count - 20} more");
-
-                TestContext.Out.WriteLine($"Extra rows (found but not expected or over-counted): {extra.Count}");
-                foreach (var l in extra.Take(20)) TestContext.Out.WriteLine($"  EXTRA:   {l}");
-                if (extra.Count > 20) TestContext.Out.WriteLine($"  ...and {extra.Count - 20} more");
-
-                Assert.Fail($"Generated top-down truncation table does not match expected.\nExpected file: {expectedPath}\nActual file: {workPath}\nMissing: {missing.Count}, Extra: {extra.Count}");
-            }
-        }
-        /// <summary>
-        /// CRITICAL: Tests MostAbundantMonoisotopicMass calculation accuracy.
-        /// Compares against Protein Prospector values. Accurate mass calculation
-        /// is essential for precursor mass matching in database searches.
-        /// </summary>
-        [Test]
-        public static void CheckMostAbundantMonoisotopicMass()
-        {
-            PeptideWithSetModifications small_pep = new PeptideWithSetModifications(new Protein("PEPTIDE", "ACCESSION"), new DigestionParams(protease: "trypsin"), 1, 7, CleavageSpecificity.Full, null, 0, new Dictionary<int, Modification>(), 0, null);
-            double small_pep_most_abundant_mass_prospector = 800.36724 - 1.0079;
-            Assert.That(small_pep.MostAbundantMonoisotopicMass, Is.EqualTo(small_pep_most_abundant_mass_prospector).Within(0.01));
-
-            PeptideWithSetModifications large_pep = new PeptideWithSetModifications(new Protein("PEPTIDEPEPTIDEPEPTIDEPEPTIDEPEPTIDEPEPTIDE", "ACCESSION"), new DigestionParams(protease: "trypsin"), 1, 42, CleavageSpecificity.Full, null, 0, new Dictionary<int, Modification>(), 0, null);
-            double large_pep_most_abundant_mass_prospector = 4709.12020 - 1.0079;
-            Assert.That(large_pep.MostAbundantMonoisotopicMass, Is.EqualTo(large_pep_most_abundant_mass_prospector).Within(0.01));
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests EssentialSequence generation for peptide grouping.
-        /// Essential sequences include only user-specified modification types and are
-        /// used for peptide-level grouping and quantification. Regression test against
-        /// expected output file ensures consistency across versions.
-        /// </summary>
-        [Test]
-        public static void TestPeptideWithSetModsEssentialSequence()
-        {
-            var psiModDeserialized = Loaders.LoadPsiMod(Path.Combine(TestContext.CurrentContext.TestDirectory, "PSI-MOD.obo2.xml"));
-            Dictionary<string, int> formalChargesDictionary = Loaders.GetFormalChargesDictionary(psiModDeserialized);
-            List<Modification> UniProtPtms = Loaders.LoadUniprot(Path.Combine(TestContext.CurrentContext.TestDirectory, "ptmlist2.txt"), formalChargesDictionary).ToList();
-
-            Dictionary<string, int> modsToWrite = new Dictionary<string, int>();
-            modsToWrite.Add("UniProt",0);
-
-            var proteinXml = ProteinDbLoader.LoadProteinXML(Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "humanGAPDH.xml"), true, DecoyType.None, UniProtPtms, false, null, out var unknownMod);
-            var gapdh = proteinXml[0];
-
-            var gapdhPeptides = gapdh.Digest(new DigestionParams(protease: "trypsin", maxMissedCleavages: 0, minPeptideLength: 1, initiatorMethionineBehavior: InitiatorMethionineBehavior.Variable), UniProtPtms, new List<Modification>());
-
-            List<string> allSequences = new List<string>();
-            foreach (var peptide in gapdhPeptides)
-            {
-                allSequences.Add(peptide.EssentialSequence(modsToWrite));
-            }
-
-            var expectedFullStrings = File.ReadAllLines(Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "essentialSequences.txt"));
-
-            CollectionAssert.AreEquivalent(expectedFullStrings, allSequences.ToArray());
-        }
-
-        [Test]
-        public static void TestPeptideWithSetModsEssentialSequence_SequenceSerializerMatchesExtension()
-        {
-            var psiModDeserialized = Loaders.LoadPsiMod(Path.Combine(TestContext.CurrentContext.TestDirectory, "PSI-MOD.obo2.xml"));
-            Dictionary<string, int> formalChargesDictionary = Loaders.GetFormalChargesDictionary(psiModDeserialized);
-            List<Modification> UniProtPtms = Loaders.LoadUniprot(Path.Combine(TestContext.CurrentContext.TestDirectory, "ptmlist2.txt"), formalChargesDictionary).ToList();
-
-            Dictionary<string, int> modsToWrite = new Dictionary<string, int>
-            {
-                { "UniProt", 0 }
-            };
-
-            var proteinXml = ProteinDbLoader.LoadProteinXML(Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "humanGAPDH.xml"), true, DecoyType.None, UniProtPtms, false, null, out var unknownMod);
-            var gapdh = proteinXml[0];
-
-            var gapdhPeptides = gapdh.Digest(new DigestionParams(protease: "trypsin", maxMissedCleavages: 0, minPeptideLength: 1, initiatorMethionineBehavior: InitiatorMethionineBehavior.Variable), UniProtPtms, new List<Modification>());
-
-            var serializer = new EssentialSequenceSerializer(modsToWrite);
-            var service = SequenceConversionService.Default;
-
-            List<string> extensionResults = new List<string>();
-            List<string> serializerResults = new List<string>();
-            List<string> serviceFullResults = new List<string>();
-            List<string> serviceStepResults = new List<string>();
-            List<string> serviceAutoFullResults = new List<string>();
-            List<string> serviceAutoStepResults = new List<string>();
-            foreach (var peptide in gapdhPeptides)
-            {
-                extensionResults.Add(peptide.EssentialSequence(modsToWrite));
-
-                var canonical = peptide.ToCanonicalSequenceBuilder().Build();
-                var serialized = serializer.Serialize(canonical);
-                serializerResults.Add(serialized ?? string.Empty);
-
-                var serviceFullConversion = service.Convert(peptide.FullSequence, "mzLib", "essential");
-                serviceFullResults.Add(serviceFullConversion ?? string.Empty);
-
-                var canon = service.Parse(peptide.FullSequence, "mzLib");
-                var serviceStepConversion = service.Serialize(canon.Value, "essential");
-                serviceStepResults.Add(serviceStepConversion ?? string.Empty);
-
-                var serviceAutoFull = service.ConvertAutoDetect(peptide.FullSequence, "essential");
-                serviceAutoFullResults.Add(serviceAutoFull ?? string.Empty);
-
-                var serviceAutoStep = service.ParseAutoDetect(peptide.FullSequence);
-                var serviceAutoStepConversion = service.Serialize(serviceAutoStep.Value, "essential");
-                serviceAutoStepResults.Add(serviceAutoStepConversion ?? string.Empty);
-            }
-
-            CollectionAssert.AreEquivalent(extensionResults, serializerResults);
-            CollectionAssert.AreEquivalent(extensionResults, serviceFullResults);
-            CollectionAssert.AreEquivalent(extensionResults, serviceStepResults);
-            CollectionAssert.AreEquivalent(extensionResults, serviceAutoFullResults);
-            CollectionAssert.AreEquivalent(extensionResults, serviceAutoStepResults);
-
-
-            var expectedFullStrings = File.ReadAllLines(Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "essentialSequences.txt"));
-            CollectionAssert.AreEquivalent(expectedFullStrings, serializerResults.ToArray());
-            CollectionAssert.AreEquivalent(expectedFullStrings, serviceFullResults.ToArray());
-            CollectionAssert.AreEquivalent(expectedFullStrings, serviceStepResults.ToArray());
-            CollectionAssert.AreEquivalent(expectedFullStrings, serviceAutoFullResults.ToArray());
-            CollectionAssert.AreEquivalent(expectedFullStrings, serviceAutoStepResults.ToArray());
-        }
-        /// <summary>
-        /// CRITICAL: Tests FullSequence and FullSequenceWithMassShift generation.
-        /// These string representations are used for peptide identification, reporting,
-        /// and spectral library matching. Regression test against expected files
-        /// ensures output format stability across versions.
-        /// </summary>
-        [Test]
-        public static void TestPeptideWithSetModsFullSequence()
-        {
-            var psiModDeserialized = Loaders.LoadPsiMod(Path.Combine(TestContext.CurrentContext.TestDirectory, "PSI-MOD.obo2.xml"));
-            Dictionary<string, int> formalChargesDictionary = Loaders.GetFormalChargesDictionary(psiModDeserialized);
-            List<Modification> UniProtPtms = Loaders.LoadUniprot(Path.Combine(TestContext.CurrentContext.TestDirectory, "ptmlist2.txt"), formalChargesDictionary).ToList(); 
-            var proteinXml = ProteinDbLoader.LoadProteinXML(Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "humanGAPDH.xml"), true, DecoyType.None, UniProtPtms, false, null, out var unknownMod);
-            var gapdh = proteinXml[0];
-
-            var gapdhPeptides = gapdh.Digest(new DigestionParams(protease: "trypsin", maxMissedCleavages: 0, minPeptideLength:1, initiatorMethionineBehavior:InitiatorMethionineBehavior.Variable),UniProtPtms,new List<Modification>());
-            
-            List<string> allSequences = new List<string>();
-            foreach (var peptide in gapdhPeptides)
-            {
-                allSequences.Add(peptide.FullSequence);
-            }
-
-            var expectedFullStrings = File.ReadAllLines(Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "fullSequences.txt"));
-            CollectionAssert.AreEquivalent(expectedFullStrings,allSequences.ToArray());
-
-            allSequences.Clear();
-            foreach (var peptide in gapdhPeptides)
-            {
-                allSequences.Add(peptide.FullSequenceWithMassShift());
-            }
-
-            var expectedFullStringsWithMassShifts = File.ReadAllLines(Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "fullSequencesWithMassShift.txt"));
-            CollectionAssert.AreEquivalent(expectedFullStringsWithMassShifts, allSequences.ToArray());
-        }
-
-        /// <summary>
-        /// EDGE CASE: Tests peptide behavior when parent protein is null or when
-        /// accessing flanking residues. PreviousResidue/NextResidue should return '-'
-        /// at protein termini or when parent is null. Important for de novo peptides.
-        /// </summary>
-        [Test]
-        public static void TestPeptideWithSetModsNoParentProtein()
-        {
-            // null parent
-            DigestionParams dParams = new DigestionParams();
-            var pwsm = new PeptideWithSetModifications("P", null,
-                digestionParams: dParams, p: null);
-            Assert.AreEqual('-', pwsm.PreviousAminoAcid);
-            Assert.AreEqual('-', pwsm.PreviousResidue);
-            Assert.AreEqual('-', pwsm.NextAminoAcid);
-            Assert.AreEqual('-', pwsm.NextResidue);
-
-            // non-null parent
-            Protein protein = new("MQLLRCFSIFSVIASVLAQELTTICEQIPSPTLESTPYSLSTTTILANGKAMQGVFEYYKSVTFVSNCGSHPSTTSKGSPINTQYVF", "P32781");
-            var pwsMods = protein.Digest(new DigestionParams(), new List<Modification>(), new List<Modification>()).ToList();
-
-            var first = pwsMods.First(p => p.BaseSequence == "MQLLRCFSIFSVIASVLAQELTTICEQIPSPTLESTPYSLSTTTILANGK");
-            Assert.AreEqual('-', first.PreviousAminoAcid);
-            Assert.AreEqual('-', first.PreviousResidue);
-            Assert.AreEqual('A', first.NextAminoAcid);
-            Assert.AreEqual('A', first.NextResidue);
-
-            var middle = pwsMods.First(p => p.BaseSequence == "SVTFVSNCGSHPSTTSK");
-            Assert.AreEqual('K', middle.PreviousAminoAcid);
-            Assert.AreEqual('K',middle.PreviousResidue);
-            Assert.AreEqual('G',middle.NextAminoAcid);
-            Assert.AreEqual('G',middle.NextResidue);
-
-            var last = pwsMods.First(p => p.BaseSequence == "GSPINTQYVF");
-            Assert.AreEqual('K', last.PreviousAminoAcid);
-            Assert.AreEqual('K', last.PreviousResidue);
-            Assert.AreEqual('-', last.NextAminoAcid);
-            Assert.AreEqual('-', last.NextResidue);
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests equality comparison for PeptideWithSetModifications.
-        /// Correct equality is essential for HashSet/Dictionary operations, peptide
-        /// deduplication, and protein inference. Tests same peptide, different proteins,
-        /// different positions, and null comparisons.
-        /// </summary>
-        [Test]
-        public static void TestPeptideWithSetModsEquals()
-        {
-            // Create two proteins
-            Protein protein1 = new Protein("SEQUENCEK", "accession1");
-            Protein protein2 = new Protein("SEQUENCEK", "accession2");
-
-            // Create digestion parameters
-            DigestionParams digestionParams = new DigestionParams(protease: "trypsin", maxMissedCleavages: 0, initiatorMethionineBehavior: InitiatorMethionineBehavior.Retain);
-
-            // Digest the proteins to get peptides
-            PeptideWithSetModifications peptide1 = protein1.Digest(digestionParams, new List<Modification>(), new List<Modification>()).First();
-            PeptideWithSetModifications peptide2 = protein2.Digest(digestionParams, new List<Modification>(), new List<Modification>()).First();
-
-            // Test equality - same peptide
-            Assert.IsTrue(peptide1.Equals(peptide1));
-
-            // different peptide
-            Assert.IsTrue(!peptide1.Equals(peptide2));
-            Assert.IsTrue(!peptide1.Equals((object)peptide2));
-            Assert.IsTrue(!peptide1.Equals((IBioPolymerWithSetMods)peptide2));
-            Assert.AreNotEqual(peptide1.GetHashCode(), peptide2.GetHashCode());
-
-            // Test inequality with different start residue
-            PeptideWithSetModifications peptide3 = new PeptideWithSetModifications(protein1, digestionParams, 2, 9, CleavageSpecificity.Full, "", 0, new Dictionary<int, Modification>(), 0);
-            Assert.IsFalse(peptide1.Equals(peptide3));
-
-            // Test inequality with different parent accession
-            PeptideWithSetModifications peptide4 = new PeptideWithSetModifications(protein2, digestionParams, 1, 9, CleavageSpecificity.Full, "", 0, new Dictionary<int, Modification>(), 0);
-            Assert.IsFalse(peptide1.Equals(peptide4));
-
-            // all fail on null
-            Assert.That(!peptide1.Equals(null));
-            Assert.That(!peptide1.Equals((object)null));
-            Assert.That(!peptide1.Equals((PeptideWithSetModifications)null));
-        }
-
-       
-
-        /// <summary>
-        /// CRITICAL: Tests modification parsing from FullSequence strings.
-        /// Validates that GetModificationDictionaryFromFullSequence and
-        /// GetModificationsFromFullSequence correctly parse modifications from
-        /// peptide strings. Essential for reading results files and spectral libraries.
-        /// </summary>
-        [Test]
-        public static void TestIBioPolymerWithSetModsModificationFromFullSequence()
-        {
-            Dictionary<string, Modification> un = new Dictionary<string, Modification>();
-            var psiModDeserialized = Loaders.LoadPsiMod(Path.Combine(TestContext.CurrentContext.TestDirectory, "PSI-MOD.obo2.xml"));
-            Dictionary<string, int> formalChargesDictionary = Loaders.GetFormalChargesDictionary(psiModDeserialized);
-            List<Modification> UniProtPtms = Loaders.LoadUniprot(Path.Combine(TestContext.CurrentContext.TestDirectory, "ptmlist2.txt"),
-                    formalChargesDictionary).ToList();
-            List<Protein> proteins = ProteinDbLoader.LoadProteinXML(Path.Combine(TestContext.CurrentContext.TestDirectory, "DatabaseTests", "cRAP_databaseGPTMD.xml"),
-                true, DecoyType.None, UniProtPtms, false, new string[] { "exclude_me" }, out un);
-            var allKnownModDict = UniProtPtms.ToDictionary(p => p.IdWithMotif, p => p);
-            var digestionParameters = new DigestionParams(maxModsForPeptides: 3);
-
-            foreach (Protein p in proteins)
-            {
-                List<PeptideWithSetModifications> digestedPeptides =
-                    p.Digest(digestionParameters, [], [], null, null).ToList();
-                // take the most modified peptide by base sequence and ensure all methods function properly
-                foreach (var targetPeptide in digestedPeptides
-                             .Where(pep => pep.FullSequence.Contains('['))
-                             .GroupBy(pep => pep.BaseSequence)
-                             .Select(pepGroup => pepGroup.MaxBy(pep => pep.AllModsOneIsNterminus.Count)))
-                {
-                    var startResidue = targetPeptide.OneBasedStartResidue;
-                    var endResidue = targetPeptide.OneBasedEndResidue;
-
-                    // Pull our expected modifications based upon parent protein object with a maximum value of DigestionParameters.MaxMods
-                    // A bunch of logic to count the number of expected modifications based upon the xml database entries
-                    int expectedModCount = 0;
-                    foreach (var modDictEntry in p.OneBasedPossibleLocalizedModifications
-                                 .Where(mod => mod.Key >= startResidue && mod.Key <= endResidue))
-                    {
-                        if (modDictEntry.Value.Count > 1)
-                        {
-                            var locRestrictions = modDictEntry.Value.Select(mod => mod.LocationRestriction).ToList();
-
-                            if (locRestrictions.AllSame())
-                            {
-                                if (locRestrictions.First() == "Anywhere.")
-                                    expectedModCount++;
-                                else if (locRestrictions.First() == "N-terminal." && modDictEntry.Key == startResidue)
-                                    expectedModCount++;
-                            }
-                            else if (modDictEntry.Value.Select(mod => mod.LocationRestriction).Contains("Anywhere.")
-                                     && modDictEntry.Value.Select(mod => mod.LocationRestriction)
-                                         .Contains("N-terminal."))
-                            {
-                                expectedModCount++;
-                                if (modDictEntry.Key == startResidue)
-                                    expectedModCount++;
-                            }
-                        }
-                        else
-                        {
-                            switch (modDictEntry.Value.First().LocationRestriction)
-                            {
-                                case "Anywhere.":
-                                case "N-terminal." when modDictEntry.Key == startResidue:
-                                    expectedModCount++;
-                                    break;
-                            }
-                        }
-                    }
-
-                    expectedModCount = Math.Min(expectedModCount, digestionParameters.MaxMods);
-
-                    var expectedModifications = p.OneBasedPossibleLocalizedModifications.Where(mod =>
-                        mod.Key >= startResidue &&
-                        mod.Key <= endResidue).SelectMany(mod => mod.Value).ToList();
-
-                    // Parse modifications from PWSM and two IBioPolymerWithSetMods methods
-                    var pwsmModDict = targetPeptide.AllModsOneIsNterminus;
-                    var bpwsmModDict = IBioPolymerWithSetMods.GetModificationDictionaryFromFullSequence(targetPeptide.FullSequence, allKnownModDict);
-                    var bpwsmModList = IBioPolymerWithSetMods.GetModificationsFromFullSequence(targetPeptide.FullSequence, allKnownModDict);
-
-                    // Ensure all methods are in agreement by modification count
-                    Assert.AreEqual(pwsmModDict.Count, expectedModCount);
-                    Assert.AreEqual(bpwsmModDict.Count, expectedModCount);
-                    Assert.AreEqual(bpwsmModList.Count, expectedModCount);
-
-                    // Ensure all methods are in agreement by modification identify
-                    foreach (var pwsmModification in pwsmModDict.Values)
-                        Assert.Contains(pwsmModification, expectedModifications);
-                    foreach (var pwsmModification in bpwsmModDict.Values)
-                        Assert.Contains(pwsmModification, expectedModifications);
-                    foreach (var pwsmModification in bpwsmModList)
-                        Assert.Contains(pwsmModification, expectedModifications);
-                }
-            }
-        }
-
-        /// <summary>
-        /// CRITICAL: Tests parsing of nucleotide substitution annotations in full sequences.
-        /// The ParseSubstitutedFullSequence method must correctly apply amino acid
-        /// substitutions from GPTMD-discovered variants while preserving other modifications.
-        /// Essential for variant peptide sequence reconstruction.
-        /// </summary>
-        [Test]
-        public static void TestGetSubstitutedFullSequence()
-        {
-            //It should take care of multiple substitutions
-            string test1 = "F[1 nucleotide substitution:F->Y on F]SIMGGGLA[1 nucleotide substitution:A->S on A]DR";
-            string expected1 = "YSIMGGGLSDR";
-            var actual1 = IBioPolymerWithSetMods.ParseSubstitutedFullSequence(test1);
-            Assert.That(actual1, Is.EqualTo(expected1));
-
-            //It should not change other modifications
-            string test2 = "SANH[1 nucleotide substitution:H->L on H]M[Common Variable:Oxidation on M]AGHWVAISGAAGGLGSLAVQYAK";
-            string expected2 = "SANLM[Common Variable:Oxidation on M]AGHWVAISGAAGGLGSLAVQYAK";
-            var actual2 = IBioPolymerWithSetMods.ParseSubstitutedFullSequence(test2);
-            Assert.That(actual2, Is.EqualTo(expected2));
-
-            //It should work on 2 nucleotide substitutions
-            string test3 = "S[2+ nucleotide substitution:S->E on S]AAADRLNLTSGHLNAGR";
-            string expected3 = "EAAADRLNLTSGHLNAGR";
-            var actual3 = IBioPolymerWithSetMods.ParseSubstitutedFullSequence(test3);
-            Assert.That(actual3, Is.EqualTo(expected3));
-        }
-    }
-}
